% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{gamljlm}
\alias{gamljlm}
\alias{gamljglm}
\alias{gamljmixed}
\alias{gamljgmixed}
\title{Alternative spelling of GAMLj main functions}
\usage{
gamljlm(...)

gamljglm(...)

gamljmixed(...)

gamljgmixed(...)
}
\arguments{
\item{...}{the same arguments that can be passed to the corresponding function}
}
\value{
a GAMLj results object equivalent to the results of (\code{\link{gamlj_lm}})

a GAMLj results object equivalent to the results of (\code{\link{gamlj_glm}})

a GAMLj results object equivalent to the results of (\code{\link{gamlj_mixed}})

a GAMLj results object equivalent to the results of (\code{\link{gamlj_gmixed}})
}
\description{
An alternative spelling of the command \code{\link{gamlj_lm}}

An alternative spelling of the command \code{\link{gamlj_glm}}

An alternative spelling of the command \code{\link{gamlj_mixed}}

An alternative spelling of the command \code{\link{gamlj_gmixed}}
}
\author{
Marcello Gallucci
}
