% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{anova}
\alias{anova}
\alias{anova.gamlj}
\title{anova test on GAMLj results}
\usage{
anova(object, ...)

\method{anova}{gamlj}(object, object2, ...)
}
\arguments{
\item{object}{a gamlj results object of the class `gamlj`}

\item{...}{all options accepted by a gamlj model function.}

\item{object2}{a gamlj results object of the class `gamlj` representing the nested model. Overiddes \code{nested_terms}.
It can be passed also as a a right-hand side formula specifying terms of the nested model.}
}
\value{
an object of class ggplot or a list of ggplot objects
}
\description{
This is a convenience function to extract the ANOVA table (omnibus tests) from a GAMLj model. If no options is passed, extracts the 
ANOVA tests table already in the model results (if any). If two GAMLj models are provided, a model comparison si produces. Any option
accepted by gamlj model can be passed.
are returned.
}
\examples{
data(fivegroups)
fivegroups$Group<-factor(fivegroups$Group)
gmod<-gamlj::gamlj_lm(
  formula = Score ~Group,
  data = fivegroups)

posthoc(gmod,formula =~Group)
}
\author{
Marcello Gallucci
}
