% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{assumptions}
\alias{assumptions}
\title{Assumptions checking plots}
\usage{
assumptions(object)
}
\arguments{
\item{object}{a gamlj results object of the class `gamlj*Results``}
}
\value{
a list of lists with titles and ggplot objects
}
\description{
This function returns a list of plots as a ggplot objects produced by the assumptions checking options, such
as `qq_plot`, `resid_plot` and `norm_plot` for continuous dependent variable models, 
`cluster_boxplot`, `cluster_respred` and `rand_hist` for mixed models.
}
\author{
Marcello Gallucci
}
