% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beers_bars}
\alias{beers_bars}
\title{Beers at bars}
\usage{
data(beers_bars)
}
\description{
Simulated data testing random coefficients regression, with smiles as dependent
variable, beers as independent variable, and bars as cluster
}
\examples{
data(beers_bars)
}
\keyword{datasets}
