% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfit.R
\name{deviance.clm}
\alias{deviance.clm}
\title{deviance function for clm models}
\usage{
\method{deviance}{clm}(object, ...)
}
\arguments{
\item{object}{an object of class  \link[ordinal]{clm}.}

\item{...}{additional optional argument.}
}
\description{
deviance function for clm models
}
