% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{em_means}
\alias{em_means}
\title{Estimated marginal means in GAMLj results}
\usage{
em_means(object, formula, ...)
}
\arguments{
\item{object}{a gamlj results object of the class `gamlj`}

\item{formula}{(optional) rhs formula defining the factor(s) for which levels expected means are estimated}

\item{...}{all options accepted by a gamlj model function.}
}
\value{
an object of class
}
\description{
This is a convenience function to extract the estimated marginal means table from a GAMLj model. If no option is passed, extracts the
emmeans table already in the model results (if any). If a formula is passed, the estimated marginal means for the defined
factors are displayed.
}
\examples{
data(fivegroups)
fivegroups$Group <- factor(fivegroups$Group)
gmod <- GAMLj3::gamlj_lm(
    formula = Score ~ Group,
    data = fivegroups
)

em_means(gmod, formula = ~Group)
}
\author{
Marcello Gallucci
}
