% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exercise}
\alias{exercise}
\title{Endurance on a threadmill.}
\format{A data frame with 245 rows and 4 variables:
\describe{
  \item{case}{subject ID}
  \item{xage}{age of the participant in years}
  \item{zexer}{years of exercising of the participant}
  \item{yendu}{endurance on the threadmill, in minutes}
  ...
}}
\source{
\url{https://books.google.it/books?hl=it&lr=&id=gkalyqTMXNEC&oi=fnd&pg=PP1&ots=tRCS--o5fh&sig=flnAHKysZzELhSgXiw8E56nIUMI}
}
\usage{
exercise
}
\description{
A dataset containing the minutes ran on a threadmill by age and years of exercise
}
\keyword{datasets}
