% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer.rsquared.R
\name{family_link.stop}
\alias{family_link.stop}
\title{stop execution if unable to calculate variance for a given family and link}
\usage{
family_link.stop(family, link)
}
\description{
stop execution if unable to calculate variance for a given family and link
}
