% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fivegroups}
\alias{fivegroups}
\title{Five groups for contrasts}
\usage{
data(fivegroups)
}
\description{
Data with five groups and a continuous dependent variable for checking contrasts results
}
\examples{
data(fivegroups)
}
\keyword{datasets}
