% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{gamlj_assumptionsPlots}
\alias{gamlj_assumptionsPlots}
\title{Assumptions checking plots}
\usage{
gamlj_assumptionsPlots(object)
}
\arguments{
\item{object}{a gamlj results object of the class `gamlj*Results``}
}
\value{
a list of lists with titles and ggplot objects
}
\description{
This function returns a list of plots as a ggplot objects produced by the assumptions checking options, such
as `residPlot`, `normPlot` (for continuous dependent variable models), `randHist` and `clusterBoxplot` (for mixed models).
}
\author{
Marcello Gallucci
}
