% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{gamlj_data}
\alias{gamlj_data}
\title{Extract data transformed by GAMLj to run R analyses in line with gamlj setup}
\usage{
gamlj_data(gobj)
}
\arguments{
\item{gobj}{a gamlj results object of the class GAMLj*}
}
\value{
a dataset
}
\description{
This function returns a dataset with the variables in the GAMLj model 
transformed according to GAMLj options. It is usefull to run additional 
models in R with other R packages with the same setup used by GAMLj
}
\examples{
data("qsport")
obj<-gamlj::gamljGlm(formula = performance ~ hours,
                data = qsport,
                scaling = c(hours="standardized"))

gdata<-gamlj_data(obj)
lm(performance ~ hours,data=gdata)
}
\author{
Marcello Gallucci
}
