% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{GAMLj_DEBUG}
\alias{GAMLj_DEBUG}
\title{Load constants and some package}
\format{An object of class \code{logical} of length 1.}
\usage{
GAMLj_DEBUG
}
\description{
Load constants and some package
}
\keyword{datasets}
