% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{gamlj_drop}
\alias{gamlj_drop}
\title{Update GAMLj results by removing a block or results}
\usage{
gamlj_drop(gobj, analysis)
}
\arguments{
\item{gobj}{a gamlj results object of the class GAMLj*}

\item{analysis}{which analysis needs to be removed. It can be `simpleEffects`, `emmans` (estimated means for factors),`plots`, or `posthoc`}
}
\value{
an object of class GAMLj* as the input object
}
\description{
This function re-estimates a GAMLj model removing some of the directives passed to the orginal model
}
\author{
Marcello Gallucci
}
