% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{gamlj_extractModel}
\alias{gamlj_extractModel}
\title{Extract the R model fro  a GAMLj object}
\usage{
gamlj_extractModel(gobj)
}
\arguments{
\item{gobj}{a gamlj results object of the class GAMLj*}
}
\value{
an R object of class lmer, lm, or glm
}
\description{
Extract the R model fro  a GAMLj object
}
\author{
Marcello Gallucci
}
