% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamljglm.f.R
\name{gamlj_glm}
\alias{gamlj_glm}
\title{Generalized Linear Model}
\usage{
gamlj_glm(
  formula = NULL,
  data,
  model_type = "linear",
  dep = NULL,
  factors = NULL,
  covs = NULL,
  offset = NULL,
  weights = NULL,
  model_terms = NULL,
  fixed_intercept = TRUE,
  nested_intercept = NULL,
  custom_family = "gaussian",
  custom_link = "identity",
  nested_terms = NULL,
  omnibus = "LRT",
  estimates_ci = FALSE,
  es_ci = FALSE,
  ci_method = "wald",
  boot_r = 1000,
  ci_width = 95,
  contrasts = NULL,
  contrast_custom_values = NULL,
  contrast_custom_focus = NULL,
  show_contrastnames = TRUE,
  show_contrastcodes = FALSE,
  vcov = FALSE,
  plot_x = NULL,
  plot_z = NULL,
  plot_terms = NULL,
  plot_by = NULL,
  plot_raw = FALSE,
  plot_yscale = FALSE,
  plot_xoriginal = FALSE,
  plot_black = FALSE,
  plot_around = "ci",
  emmeans = NULL,
  posthoc = NULL,
  simple_x = NULL,
  simple_mods = NULL,
  simple_interactions = FALSE,
  covs_conditioning = "mean_sd",
  ccm_value = 1,
  ccp_value = 25,
  covs_scale_labels = "labels",
  adjust = list("bonf"),
  covs_scale = NULL,
  expb_ci = TRUE,
  es = list("expb", "eta"),
  propodds_test = FALSE,
  plot_scale = "response",
  se_method = "standard"
)
}
\arguments{
\item{formula}{(optional) the formula of the model compatible with \link[stats]{glm}. If not passed,
model terms should be defined as a list in  \code{model_terms} option.
For logistic and probit models formula can be of the form `cbind(success,fail) ~ x ` where
`success` is the counts of success (y=1) and fail is the counts of failure (y=0). It accepts
also the the form `probabilities / totals ~ x ` where `probabilities` is the probability of success
and `totals` is the number of cases.}

\item{data}{the data as a data.frame}

\item{model_type}{define the type of model (link function and distribution combination) required.
It can be \code{linear}, \code{poisson}, \code{poiover}, \code{nb},
\code{logistic}, \code{probit}, \code{custom}, \code{ordinal}, \code{multinomial}, and \code{beta}.}

\item{dep}{a string naming the dependent variable from \code{data}; the
variable must be numeric. Not needed if \code{formula} is used.}

\item{factors}{a vector of strings naming the fixed factors from
\code{data}. Not needed if \code{formula} is used.}

\item{covs}{a vector of strings naming the covariates from \code{data}. Not
needed if \code{formula} is used.}

\item{offset}{a vector of strings naming the offset variables.}

\item{weights}{an optional variable name indicating the vector of ‘prior weights’ to be used in the fitting process. Should be NULL or a variable name in the data.}

\item{model_terms}{a list of character vectors describing fixed effects
terms. Not needed if \code{formula} is used.}

\item{fixed_intercept}{\code{TRUE} (default) or \code{FALSE}, estimates
fixed intercept. Not needed if \code{formula} is used.}

\item{nested_intercept}{\code{TRUE} (default) or \code{FALSE}, estimates
fixed intercept. Not needed if \code{formula} is used.}

\item{custom_family}{Distribution family for the custom model, accepts
\code{gaussian}, \code{binomial}, \code{gamma} and \code{inverse_gaussian}.}

\item{custom_link}{Distribution family for the \code{model_type="custom"}, accepts
\code{identity}, \code{log} and \code{inverse}. Use \code{onemu2}  for 1/mu^2).}

\item{nested_terms}{a list of character vectors describing effects terms
for nested model. It can be passed as right-hand formula.}

\item{omnibus}{Whether the omnibus test for the model should be \code{wald} or \code{LRT}.}

\item{estimates_ci}{\code{TRUE} (default) or \code{FALSE} , coefficients CI
in tables}

\item{es_ci}{\code{TRUE} or \code{FALSE} (default), effect size indices
confidence intervals in the effect size table}

\item{ci_method}{The method used to compute the confidence intervals. `wald` uses the Wald method to compute standard
errors and confidence intervals. `profile` computes Profile Likelihood Based Confidence Interval, in which
the bounds are chosen based on the percentiles of the chi-square distribution around the maximum likelihood
estimate. `quantile` performs a non-parametric boostrap, with `boot_r` repetitions, and compute
the CI based on the percentiles of the boostrap distribution. `bcai` implements the bias-corrected bootstrap method.}

\item{boot_r}{a number bootstrap repetitions.}

\item{ci_width}{a number between 50 and 99.9 (default: 95) specifying the
confidence interval width for the plots.}

\item{contrasts}{a named vector of the form \code{c(var1="type",
var2="type2")} specifying the type of contrast to use, one of
\code{deviation}, \code{simple}, \code{dummy}, \code{difference},
\code{helmert}, \code{repeated} or \code{'polynomial'}. If NULL,
\code{simple} is used. Can also be passed as a list of list of the form
list(list(var="var1",type="type1")).}

\item{contrast_custom_values}{a named list with the custom contrast weights, of the form \code{list(factorname=numeric vector)}, for instance \code{list(factorname=c(1,1,-2))}.
only one constrast per variable is allowed.}

\item{contrast_custom_focus}{if any factor is coded with \code{'custom'}, when \code{TRUE} or \code{NULL } (default) the coefficients, simple effects and simple interactions
are focused on the custom contrast. If \code{FALSE} , variables are coded accordingly to the passed contrast, but
no special table or test is devoted to the contrast.}

\item{show_contrastnames}{\code{TRUE} or \code{FALSE} (default), shows raw
names of the contrasts variables in tables}

\item{show_contrastcodes}{\code{TRUE} or \code{FALSE} (default), shows
contrast coefficients tables}

\item{vcov}{\code{TRUE} or \code{FALSE} (default), shows coefficients
covariances}

\item{plot_x}{a string naming the variable placed on the horizontal axis of
the plot}

\item{plot_z}{a string naming the variable represented as separate lines in
the plot}

\item{plot_terms}{a RHS formulas. It is an alternative way to
produce plots. \code{plot_terms=~x+z:x} produces two plots, one with
`x` as x-axis and one with z as x-axis and x defining different lines.}

\item{plot_by}{a list of variables defining the levels at which a separate
plot should be produced.}

\item{plot_raw}{\code{TRUE} or \code{FALSE} (default), plot raw data along
the predicted values}

\item{plot_yscale}{\code{TRUE} or \code{FALSE} (default), set the Y-axis
range equal to the range of the observed values.}

\item{plot_xoriginal}{\code{TRUE} or \code{FALSE} (default), use original
scale for covariates.}

\item{plot_black}{\code{TRUE} or \code{FALSE} (default), use different
linetypes per levels.}

\item{plot_around}{\code{'none'} (default), \code{'ci'}, or \code{'se'}.
Use no error bars, use confidence intervals, or use standard errors on the
plots, respectively.}

\item{emmeans}{a rhs formula with the terms specifying the marginal means
to estimate (of the form \code{'~x+x:z'})}

\item{posthoc}{a rhs formula with the terms specifying the table to apply
the comparisons (of the form \code{'~x+x:z'}). The formula is not expanded,
so \code{x*z} becomes \code{x+z} and not \code{x+z+x:z}. It can be
passed also as a list of the form \code{'list("x","z",c("x","z")'}}

\item{simple_x}{The variable for which the simple effects (slopes)
are computed}

\item{simple_mods}{the variable that provides the levels at which the
simple effects are computed}

\item{simple_interactions}{should simple Interactions be computed}

\item{covs_conditioning}{\code{'mean_sd'} (default), \code{'custom'} , or
\code{'percent'}. Use to condition the covariates (if any)}

\item{ccm_value}{how many st.deviations around the means used to condition
simple effects and plots. Used if \code{simpleScale}=\code{'mean_sd'}}

\item{ccp_value}{offsett (number of percentiles) around the median used to
condition simple effects and plots. Used if
\code{simpleScale}=\code{'percent'}}

\item{covs_scale_labels}{how the levels of a continuous moderator should
appear in tables and plots: \code{labels}, \code{values} and
\code{values_labels}, \code{ovalues}, \code{ovalues_labels}. The latter two refer
to the variable original levels, before scaling.}

\item{adjust}{one or more of \code{'none'},  \code{'bonf'},\code{'tukey'}
\code{'holm'}; provide no,  Bonferroni, Tukey and Holm Post Hoc corrections
respectively.}

\item{covs_scale}{a named vector of the form \code{c(var1='type',
var2='type2')} specifying the transformation to apply to covariates, one of
\code{'centered'} to the mean, \code{'standardized'},\code{'log'} or
\code{'none'}. \code{'none'} leaves the variable as it is.}

\item{expb_ci}{\code{TRUE} (default) or \code{FALSE} , exp(B) CI in table}

\item{es}{Effect size indices. \code{expb} (default) exponentiates the
coefficients. For dichotomous dependent variables relative risk indices
(RR) can be obtained. \code{marginals} computes the marginal effects.
\code{eta} and \code{epsilon} request generalized eta-squared and
epsilon-squared indices for the ANOVA effects.}

\item{propodds_test}{Test parallel lines assumptions in cumulative link
model (ordinal regression)}

\item{plot_scale}{Chi-squared computation method. \code{'lrt'} (default)
gives LogLikelihood ration test,  \code{'wald'} gives the Wald Chi-squared.}

\item{se_method}{Method to compute the standard error.
Classical standard errors is the default \code{standard}.
Four methods for  heteroschedasticy-consistent
standard errors are available: \code{HC0},
\code{HC1},\code{HC2},\code{HC3}, from package \code{sandwich} .
See \code{\link[sandwich]{vcovHC}} for details.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$model} \tab \tab \tab \tab \tab The underlying estimated model \cr
  \code{results$info} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$r2} \tab \tab \tab \tab \tab a table of R \cr
  \code{results$main$fit} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$anova} \tab \tab \tab \tab \tab a table of ANOVA results \cr
  \code{results$main$coefficients} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$vcov} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$contrastCodeTables} \tab \tab \tab \tab \tab an array of contrast coefficients tables \cr
  \code{results$main$marginals} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$relativerisk} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$paralleltest} \tab \tab \tab \tab \tab a table \cr
  \code{results$posthoc} \tab \tab \tab \tab \tab an array of post-hoc tables \cr
  \code{results$simpleEffects$anova} \tab \tab \tab \tab \tab a table of ANOVA for simple effects \cr
  \code{results$simpleEffects$coefficients} \tab \tab \tab \tab \tab a table \cr
  \code{results$simpleInteractions} \tab \tab \tab \tab \tab an array of simple interactions tables \cr
  \code{results$emmeans} \tab \tab \tab \tab \tab an array of predicted means tables \cr
  \code{results$mainPlots} \tab \tab \tab \tab \tab an array of results plots \cr
  \code{results$plotnotes} \tab \tab \tab \tab \tab a html \cr
  \code{results$predicted} \tab \tab \tab \tab \tab an output \cr
  \code{results$residuals} \tab \tab \tab \tab \tab an output \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$info$asDF}

\code{as.data.frame(results$info)}
}
\description{
Generalized Linear Model with options to estimate logistic, probit,
ordinal, multinomial and custum link function models. Provides options
to estimate posthoc, simple effects and slopes, plots of main effects and
interaction. Model comparison is also an option.
}
\examples{
data <- emmeans::neuralgia
GAMLj3::gamlj_glm(
    formula = Pain ~ Duration,
    data = data,
    model_type = "logistic"
)

}
