% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamljlm.f.R
\name{gamlj_lm}
\alias{gamlj_lm}
\title{General Linear Model}
\usage{
gamlj_lm(
  formula = NULL,
  data,
  dep = NULL,
  fixed_intercept = TRUE,
  factors = NULL,
  covs = NULL,
  model_terms = NULL,
  nested_terms = NULL,
  nested_intercept = NULL,
  omnibus = "F",
  estimates_ci = TRUE,
  es_ci = FALSE,
  betas_ci = FALSE,
  ci_width = 95,
  ci_method = "wald",
  boot_r = 1000,
  contrasts = NULL,
  contrast_custom_values = NULL,
  contrast_custom_focus = NULL,
  show_contrastnames = TRUE,
  show_contrastcodes = FALSE,
  vcov = FALSE,
  plot_x = NULL,
  plot_z = NULL,
  plot_by = NULL,
  plot_terms = NULL,
  plot_raw = FALSE,
  plot_yscale = FALSE,
  plot_xoriginal = FALSE,
  plot_black = FALSE,
  plot_around = "ci",
  plot_jn = FALSE,
  emmeans = NULL,
  posthoc = NULL,
  simple_x = NULL,
  simple_mods = NULL,
  simple_interactions = FALSE,
  covs_scale = NULL,
  covs_conditioning = "mean_sd",
  ccm_value = 1,
  ccp_value = 25,
  ccra_steps = 3,
  covs_scale_labels = "labels",
  adjust = list("bonf"),
  posthoc_es = list("dm"),
  d_ci = FALSE,
  es = list("beta", "etap"),
  homo_test = FALSE,
  colli_test = FALSE,
  qq_plot = FALSE,
  norm_test = FALSE,
  norm_plot = FALSE,
  resid_plot = FALSE,
  intercept_info = FALSE,
  es_info = FALSE,
  dep_scale = "none",
  se_method = "standard"
)
}
\arguments{
\item{formula}{(optional) the formula of the model, see the examples. If not passed
model terms should be defined as a list in  \code{model_terms} option.}

\item{data}{the data as a data frame}

\item{dep}{a string naming the dependent variable from \code{data}; the
variable must be numeric. Not needed if \code{formula} is used.}

\item{fixed_intercept}{\code{TRUE} (default) or \code{FALSE}, estimates
fixed intercept. Overridden if \code{formula} is used and contains \code{~1} or \code{~0}.}

\item{factors}{a vector of strings naming the fixed factors from
\code{data}. Not needed if \code{formula} is used.}

\item{covs}{a vector of strings naming the covariates from \code{data}. Not
needed if \code{formula} is used.}

\item{model_terms}{a list of character vectors describing fixed effects
terms. Not needed if \code{formula} is used.}

\item{nested_terms}{A right-hand formula for the nested model. It can be passed as a list of character vectors describing effects terms
for the nested model.}

\item{nested_intercept}{\code{TRUE} (default) or \code{FALSE}, estimates
fixed intercept. Overridden if \code{model_ters} is a formula and contains \code{~1} or \code{~0}..}

\item{omnibus}{Omnibus tests are based on F-test \code{F} (default) or
loglikelihood ration test \code{LRT}.}

\item{estimates_ci}{\code{TRUE} (default) or \code{FALSE} , parameters CI
in table}

\item{es_ci}{\code{TRUE} or \code{FALSE} (default), effect size indices
confidence intervals in the effect size table}

\item{betas_ci}{\code{TRUE} (default) or \code{FALSE} , parameters CI in
table}

\item{ci_width}{a number between 50 and 99.9 (default: 95) specifying the
confidence interval width for the plots.}

\item{ci_method}{the method to compute the confidence intervals. It can be `wald` (default) for
large samples confidence intervals, `quantile` for percentile bootstrap method, or `bcai`
for bias corrected accelarated method.}

\item{boot_r}{a number bootstrap repetitions.}

\item{contrasts}{a named vector of the form \code{c(var1="type",
var2="type2")} specifying the type of contrast to use, one of
\code{'deviation'}, \code{'simple'}, \code{'dummy'}, \code{'difference'},
\code{'helmert'}, \code{'repeated'}, \code{'polynomial'} \code{'custom'}. If NULL,
\code{simple} is used. Can also be passed as a list of list of the form
list(list(var="var1",type="type1")). If any factor contrast is defined as \code{'custom'},
a named list of the form \code{list(factorname=numeric vector)}, for instance \code{list(factorname=c(1,1,-2))} should be passed
to the option \code{contrast_custom_values}.}

\item{contrast_custom_values}{a named list with the custom contrast weights, of the form \code{list(factorname=numeric vector)}, for instance \code{list(factorname=c(1,1,-2))}.
only one constrast per variable is allowed.}

\item{contrast_custom_focus}{if any factor is coded with \code{'custom'}, when \code{TRUE} or \code{NULL } (default) the coefficients, simple effects and simple interactions
are focused on the custom contrast. If \code{FALSE} , variables are coded accordingly to the passed contrast, but
no special table or test is devoted to the contrast.}

\item{show_contrastnames}{\code{TRUE} or \code{FALSE} (default), shows raw
names of the contrasts variables in tables}

\item{show_contrastcodes}{\code{TRUE} or \code{FALSE} (default), shows
contrast coefficients tables}

\item{vcov}{\code{TRUE} or \code{FALSE} (default), shows coefficients
covariances}

\item{plot_x}{a string naming the variable placed on the horizontal axis of
the plot}

\item{plot_z}{a string naming the variable represented as separate lines in
the plot}

\item{plot_by}{a list of variables defining the levels at which a separate
plot should be produced.}

\item{plot_terms}{a RHS formulas. It is an alternative way to
produce plots. \code{plot_terms=~x+z:x} produces two plots, one with
`x` as x-axis and one with z as x-axis and x defining different lines.}

\item{plot_raw}{\code{TRUE} or \code{FALSE} (default), plot raw data along
the predicted values}

\item{plot_yscale}{\code{TRUE} or \code{FALSE} (default), set the Y-axis
range equal to the range of the observed values.}

\item{plot_xoriginal}{\code{TRUE} or \code{FALSE} (default), use original
scale for covariates.}

\item{plot_black}{\code{TRUE} or \code{FALSE} (default), use different
linetypes per levels.}

\item{plot_around}{\code{'none'} (default), \code{'ci'}, or \code{'se'}.
Use no error bars, use confidence intervals, or use standard errors on the
plots, respectively.}

\item{plot_jn}{\code{'FALSE'} (default), \code{'TRUE'}.
produces Johnson-Neyman plots for the variables specified in the main plot.}

\item{emmeans}{a rhs formula with the terms specifying the marginal means
to estimate (of the form \code{'~x+x:z'})}

\item{posthoc}{a rhs formula with the terms specifying the table to apply
the comparisons (of the form \code{'~x+x:z'}). The formula is not expanded,
so \code{'x*z'} becomes \code{'x+z'} and not \code{'x+z+x:z'}. It can be
passed also as a list of the form \code{list("x","z",c("x","z")}}

\item{simple_x}{The variable for which the simple effects (slopes)
are computed}

\item{simple_mods}{a character vector with the variable(s) providing the levels at which the
simple effects are computed}

\item{simple_interactions}{should simple Interactions be computed}

\item{covs_scale}{a named vector of the form \code{c(var1='type',
var2='type2')} specifying the transformation to apply to covariates, one of
\code{'centered'} to the mean, \code{'standardized'} or  \code{'none'}.
\code{'none'} leaves the variable as it is.}

\item{covs_conditioning}{'\code{mean_sd}' (default), or '\code{percent}'.
How to condition covariates in simple effects and plots. '\code{mean_sd}'
for mean +/- '\code{ccp_value}' * sd. '\code{percent}' for median
+/-'\code{ccp_value}' for percentiles. \code{range} conditions to min and max divided in \code{ccra_steps} steps (aproximately).}

\item{ccm_value}{how many st.deviations around the means used to condition
simple effects and plots. Used if \code{covs_conditioning}=\code{'mean_sd'}}

\item{ccp_value}{offsett (number of percentiles) around the median used to
condition simple effects and plots. Used if
\code{simpleScale}=\code{'percent'}}

\item{ccra_steps}{Covariate condition  steps from min to max:
At how many values between min and max should the covariate be conditioned \code{simpleScale}=\code{'range'}}

\item{covs_scale_labels}{how the levels of a continuous moderator should
appear in tables and plots: \code{labels}, \code{values} and
\code{values_labels}, \code{ovalues}, `ovalues_labels. The latter two refer
to the variable orginal levels, before scaling.}

\item{adjust}{adjustment method for postho tests. One or more of
\code{'none'},  \code{'bonf'},\code{'tukey'}  \code{'holm'}; provide no,
Bonferroni, Tukey and Holm Post Hoc corrections respectively.}

\item{posthoc_es}{effect size indices for mean comparisons. One or more of
\code{'dm'},  \code{'ds'},\code{'g'} for Cohen's d (dm=model SD,ds=sample
SD )  or Hedge's g}

\item{d_ci}{\code{TRUE} or \code{FALSE} (default), d confidence intervals}

\item{es}{a list of effect sizes to print out. They can be:  \code{"eta"}
for eta-squared, \code{'etap'} for partial eta-squared, \code{'omega'} for
omega-squared, \code{'omegap'} for partial omega-squared, \code{'epsilon'}
for epsilon-squared, \code{'epsilonp'} for partial epsilon-squared  and
\code{'beta'} for standardized  coefficients (betas). Default is
\code{"beta"} and \code{"parEta"}.}

\item{homo_test}{\code{TRUE} or \code{FALSE} (default), performs homogeneity
tests}

\item{colli_test}{\code{TRUE} or \code{FALSE} (default), computes VIF and Tollerance
for the terms in the model}

\item{qq_plot}{\code{TRUE} or \code{FALSE} (default), provide a Q-Q plot of
residuals}

\item{norm_test}{\code{TRUE} or \code{FALSE} (default), provide a test for
normality of residuals}

\item{norm_plot}{\code{TRUE} or \code{FALSE} (default), provide a histogram
of residuals superimposed by a normal distribution}

\item{resid_plot}{\code{TRUE} or \code{FALSE} (default), provide a
scatterplot of the residuals against predicted}

\item{intercept_info}{\code{TRUE} or \code{FALSE} (default), provide
ìnformation about the intercept (F test, effect size indexes)}

\item{es_info}{\code{TRUE} or \code{FALSE} (default), provide ìnformation
about the effect size indexes}

\item{dep_scale}{Re-scale the dependent variable.}

\item{se_method}{Method to compute the standard error.
Classical standard errors is the default \code{standard}.
Four methods for  heteroschedasticy-consistent
standard errors are available: \code{HC0},
\code{HC1},\code{HC2},\code{HC3}, from package \code{sandwich} .
See \code{\link[sandwich]{vcovHC}} for details.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$model} \tab \tab \tab \tab \tab a property \cr
  \code{results$info} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$r2} \tab \tab \tab \tab \tab a table of R \cr
  \code{results$main$intercept} \tab \tab \tab \tab \tab a table of information for the model intercept \cr
  \code{results$main$anova} \tab \tab \tab \tab \tab a table of ANOVA results \cr
  \code{results$main$effectsizes} \tab \tab \tab \tab \tab a table of effect size indeces \cr
  \code{results$main$coefficients} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$vcov} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$contrastCodeTables} \tab \tab \tab \tab \tab an array of contrast coefficients tables \cr
  \code{results$posthoc} \tab \tab \tab \tab \tab an array of post-hoc tables \cr
  \code{results$posthocEffectSize} \tab \tab \tab \tab \tab an array of post-hoc effect size \cr
  \code{results$simpleEffects$anova} \tab \tab \tab \tab \tab a table of ANOVA for simple effects \cr
  \code{results$simpleEffects$coefficients} \tab \tab \tab \tab \tab a table \cr
  \code{results$simpleInteractions} \tab \tab \tab \tab \tab an array of simple interactions tables \cr
  \code{results$emmeans} \tab \tab \tab \tab \tab an array of predicted means tables \cr
  \code{results$mainPlots} \tab \tab \tab \tab \tab an array of results plots \cr
  \code{results$plotnotes} \tab \tab \tab \tab \tab a html \cr
  \code{results$assumptions$homotest} \tab \tab \tab \tab \tab a table of homogeneity tests \cr
  \code{results$assumptions$normtest} \tab \tab \tab \tab \tab a table of normality tests \cr
  \code{results$assumptions$qqplot} \tab \tab \tab \tab \tab a q-q plot \cr
  \code{results$assumptions$normPlot} \tab \tab \tab \tab \tab Residual histogram \cr
  \code{results$assumptions$residPlot} \tab \tab \tab \tab \tab Residual Predicted plot \cr
  \code{results$predicted} \tab \tab \tab \tab \tab an output \cr
  \code{results$residuals} \tab \tab \tab \tab \tab an output \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$info$asDF}

\code{as.data.frame(results$info)}
}
\description{
General Linear Model. Estimates models using \code{lm()} function and
provides options to facilitate estimation of
interactions, simple slopes, simple effects, post-hoc tests, contrast
analysis, effect size indexes and visualization of the results.
}
\examples{
data("ToothGrowth")
GAMLj3::gamlj_lm(formula = len ~ supp, data = ToothGrowth)
}
