% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamljmixed.f.R
\name{gamlj_mixed}
\alias{gamlj_mixed}
\title{Mixed Model}
\usage{
gamlj_mixed(
  formula = NULL,
  data,
  dep = NULL,
  factors = NULL,
  covs = NULL,
  model_terms = NULL,
  fixed_intercept = TRUE,
  nested_terms = NULL,
  nested_intercept = NULL,
  omnibus = "LRT",
  estimates_ci = TRUE,
  ci_method = "wald",
  boot_r = 1000,
  ci_width = 95,
  contrasts = NULL,
  show_contrastnames = TRUE,
  show_contrastcodes = FALSE,
  plot_x = NULL,
  plot_z = NULL,
  plot_by = NULL,
  plot_raw = FALSE,
  plot_yscale = FALSE,
  plot_xoriginal = FALSE,
  plot_black = FALSE,
  plot_around = "none",
  plot_re = FALSE,
  plot_re_method = "average",
  emmeans = NULL,
  posthoc = NULL,
  simple_x = NULL,
  simple_mods = NULL,
  simple_interactions = FALSE,
  covs_conditioning = "mean_sd",
  ccm_value = 1,
  ccp_value = 25,
  covs_scale_labels = "labels",
  adjust = list("bonf"),
  covs_scale = NULL,
  dep_scale = "none",
  scale_missing = "complete",
  norm_test = FALSE,
  cluster = NULL,
  re = NULL,
  nested_re = NULL,
  reml = TRUE,
  re_lrt = FALSE,
  re_ci = FALSE,
  df_method = "Satterthwaite",
  norm_plot = FALSE,
  qq_plot = FALSE,
  resid_plot = FALSE,
  cluster_boxplot = FALSE,
  cluster_respred = FALSE,
  rand_hist = FALSE
)
}
\arguments{
\item{formula}{(optional) the formula of the linear mixed model as defined in \link[lme4]{lmer}.}

\item{data}{the data as a data frame}

\item{dep}{a string naming the dependent variable from \code{data}; the
variable must be numeric. Not needed if \code{formula} is used.}

\item{factors}{a vector of strings naming the fixed factors from
\code{data}. Not needed if \code{formula} is used.}

\item{covs}{a vector of strings naming the covariates from \code{data}. Not
needed if \code{formula} is used.}

\item{model_terms}{a list of character vectors describing fixed effects
terms. Not needed if \code{formula} is used.}

\item{fixed_intercept}{\code{TRUE} (default) or \code{FALSE}, estimates
fixed intercept. Overridden if \code{formula} is used and contains \code{~1} or \code{~0}.}

\item{nested_terms}{a list of character vectors describing effects terms
for nestet. It can be passed as right-hand formula.}

\item{nested_intercept}{\code{TRUE} (default) or \code{FALSE}, estimates
fixed intercept. Not needed if \code{formula} is used.}

\item{omnibus}{\code{TRUE} (default) or \code{FALSE}, estimates fixed
intercept. Not needed if \code{formula} is used.}

\item{estimates_ci}{\code{TRUE} (default) or \code{FALSE} , parameters CI
in table}

\item{ci_method}{.}

\item{boot_r}{a number bootstrap repetitions.}

\item{ci_width}{a number between 50 and 99.9 (default: 95) specifying the
confidence interval width for the plots.}

\item{contrasts}{a named vector of the form \code{c(var1="type",
var2="type2")} specifying the type of contrast to use, one of
\code{'deviation'}, \code{'simple'}, \code{'dummy'}, \code{'difference'},
\code{'helmert'}, \code{'repeated'} or \code{'polynomial'}. If NULL,
\code{simple} is used. Can also be passed as a list of list of the form
list(list(var="var1",type="type1")).}

\item{show_contrastnames}{\code{TRUE} or \code{FALSE} (default), shows raw
names of the contrasts variables in tables}

\item{show_contrastcodes}{\code{TRUE} or \code{FALSE} (default), shows
contrast coefficients tables}

\item{plot_x}{a string naming the variable placed on the horizontal axis of
the plot}

\item{plot_z}{a string naming the variable represented as separate lines in
the plot}

\item{plot_by}{a list of string naming the variables defining the levels
for multiple plots}

\item{plot_raw}{\code{TRUE} or \code{FALSE} (default), plot raw data along
the predicted values}

\item{plot_yscale}{\code{TRUE} or \code{FALSE} (default), set the Y-axis
range equal to the range of the observed values.}

\item{plot_xoriginal}{\code{TRUE} or \code{FALSE} (default), use original
scale for covariates.}

\item{plot_black}{\code{TRUE} or \code{FALSE} (default), use different
line type per levels.}

\item{plot_around}{\code{'none'} (default), \code{'ci'}, or \code{'se'}.
Use no error bars, use confidence intervals, or use standard errors on the
plots, respectively.}

\item{plot_re}{\code{TRUE} or \code{FALSE} (default), add predicted values
based on random effect in plot}

\item{plot_re_method}{}

\item{emmeans}{a rhs formula with the terms specifying the marginal means
to estimate (of the form \code{'~x+x:z'})}

\item{posthoc}{a rhs formula with the terms specifying the table to apply
the comparisons (of the form \code{'~x+x:z'}). The formula is not expanded,
so '\code{x*z}' becomes '\code{x+z' and not '}x+z+x:z\code{'. It can be
passed also as a list of the form '}list("x","z",c("x","z")`'}

\item{simple_x}{The variable for which the simple effects (slopes)
are computed}

\item{simple_mods}{the variable that provides the levels at which the
simple effects are computed}

\item{simple_interactions}{should simple Interactions be computed}

\item{covs_conditioning}{\code{'mean_sd'} (default), \code{'custom'} , or
\code{'percent'}. Use to condition the covariates (if any)}

\item{ccm_value}{Covariates conditioning mean offset value: how many
st.deviations around the means used to condition simple effects and plots.
Used if \code{covs_conditioning}=\code{'mean_sd'}}

\item{ccp_value}{Covariates conditioning percentile offset value: number of
percentiles around the median used to condition simple effects and plots.
Used if \code{covs_conditioning}=\code{'percent'}}

\item{covs_scale_labels}{how the levels of a continuous moderator should
appear in tables and plots: \code{labels}, \code{values} and
\code{values_labels}, \code{ovalues}, `ovalues_labels. The latter two refer
to the variable orginal levels, before scaling.}

\item{adjust}{one or more of \code{'none'},  \code{'bonf'},\code{'tukey'}
\code{'holm'},\code{'scheffe'}, \code{'tukey'}; provide no,  Bonferroni,
Tukey and Holm Post Hoc corrections respectively.}

\item{covs_scale}{a list of lists specifying the covariates scaling, one of
\code{'centered to the mean'}, \code{'standardized'}, or \code{'none'}.
\code{'none'} leaves the variable as it is}

\item{dep_scale}{Re-scale the dependent variable.}

\item{scale_missing}{.}

\item{norm_test}{\code{TRUE} or \code{FALSE} (default), provide a test for
normality of residuals}

\item{cluster}{a vector of strings naming the clustering variables from
\code{data}}

\item{re}{a list of lists specifying the models random effects.}

\item{nested_re}{a list of lists specifying the models random effects.}

\item{reml}{\code{TRUE} (default) or \code{FALSE}, should the Restricted ML
be used rather than ML}

\item{re_lrt}{\code{TRUE} or \code{FALSE} (default), LRT for the random
effects}

\item{re_ci}{\code{TRUE} or \code{FALSE} (default), confidence intervals
for the random effects}

\item{df_method}{The method for computing the denominator degrees of
freedom and F-statistics. "Satterthwaite" (default) uses Satterthwaite’s
method; "Kenward-Roger" uses Kenward-Roger’s method, "lme4" returns the
lme4-anova table, i.e., using the anova method for lmerMod objects as
defined in the lme4-package}

\item{norm_plot}{\code{TRUE} or \code{FALSE} (default), provide a histogram
of residuals superimposed by a normal distribution}

\item{qq_plot}{\code{TRUE} or \code{FALSE} (default), provide a Q-Q plot of
residuals}

\item{resid_plot}{\code{TRUE} or \code{FALSE} (default), provide a
scatterplot of the residuals against predicted}

\item{cluster_boxplot}{\code{TRUE} or \code{FALSE} (default), provide a
boxplot of random effects by the first defined clustering variable}

\item{cluster_respred}{\code{TRUE} or \code{FALSE} (default), residuals vs
predicted by the first defined clustering variable}

\item{rand_hist}{\code{TRUE} or \code{FALSE} (default), provide histogram
of random Coefficients}

\item{re_corr}{\code{'all'}, \code{'none'} (default), or \code{'block'}.
When random effects are passed as list of length 1, it decides whether the
effects should be correlated,  non correlated. If \code{'re'} is a list of
lists of length > 1, the option is automatially set to \code{'block'}. The
option is ignored if the model is passed using \code{formula}.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$model} \tab \tab \tab \tab \tab a property \cr
  \code{results$info} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$r2} \tab \tab \tab \tab \tab a table of R \cr
  \code{results$main$anova} \tab \tab \tab \tab \tab a table of ANOVA results \cr
  \code{results$main$coefficients} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$contrastCodeTables} \tab \tab \tab \tab \tab an array of contrast coefficients tables \cr
  \code{results$main$random} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$randomcov} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$ranova} \tab \tab \tab \tab \tab a table \cr
  \code{results$posthoc} \tab \tab \tab \tab \tab an array of post-hoc tables \cr
  \code{results$simpleEffects$anova} \tab \tab \tab \tab \tab a table of ANOVA for simple effects \cr
  \code{results$simpleEffects$coefficients} \tab \tab \tab \tab \tab a table \cr
  \code{results$simpleInteractions} \tab \tab \tab \tab \tab an array of simple interactions tables \cr
  \code{results$emmeans} \tab \tab \tab \tab \tab an array of predicted means tables \cr
  \code{results$mainPlots} \tab \tab \tab \tab \tab an array of results plots \cr
  \code{results$plotnotes} \tab \tab \tab \tab \tab a html \cr
  \code{results$assumptions$normtest} \tab \tab \tab \tab \tab a table of normality tests \cr
  \code{results$assumptions$qqplot} \tab \tab \tab \tab \tab a q-q plot \cr
  \code{results$assumptions$normplot} \tab \tab \tab \tab \tab Residual histogram \cr
  \code{results$assumptions$residPlot} \tab \tab \tab \tab \tab Residual Predicted plot \cr
  \code{results$assumptions$clusterBoxplot} \tab \tab \tab \tab \tab Residuals boxplot by cluster \cr
  \code{results$assumptions$clusterResPred} \tab \tab \tab \tab \tab an array of random coefficients histograms \cr
  \code{results$assumptions$randHist} \tab \tab \tab \tab \tab an array of random coefficients histograms \cr
  \code{results$predicted} \tab \tab \tab \tab \tab an output \cr
  \code{results$residuals} \tab \tab \tab \tab \tab an output \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$info$asDF}

\code{as.data.frame(results$info)}
}
\description{
Mixed Linear Model. Estimates models using \link[lme4]{lmer} and \link[lmerTests]{merTest-package} functions and 
provides options to facilitate estimation of 
interactions, simple slopes, simple effects, post-hoc tests, contrast 
analysis, effect size indexes and visualization of the results.
}
\examples{
data(subjects_by_stimuli)
gamlj::gamljMixed(
       formula = y ~ 1 + cond+( 1|subj ),
       data = subjects_by_stimuli)
}
