% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{gamlj_model}
\alias{gamlj_model}
\title{Extract the R model object contained in the GAMLj results}
\usage{
gamlj_model(gobj)
}
\arguments{
\item{gobj}{a gamlj results object of the class GAMLj*}
}
\value{
a R object from lm(), glm(), lmer() or glmer()
}
\description{
This function returns the R model  object. It is usefull to run additional 
analysis on model results.
}
\examples{
data("qsport")
obj<-gamlj::gamljGlm(
   formula = performance ~ hours,
   data = qsport)
 model<-gamlj_model(obj)
 
}
\author{
Marcello Gallucci
}
