% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{gamlj_options}
\alias{gamlj_options}
\title{GAMLj models options}
\usage{
gamlj_options(opt, value)
}
\arguments{
\item{opt}{a value for the corresponding option}
}
\description{
This function sets GAMLj suite general options
}
\details{
Options are: debug (TRUE or FALSE). When TRUE some additional warnings and info are printed in output
}
\author{
Marcello Gallucci
}
