% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{gamlj_plot}
\alias{gamlj_plot}
\title{Update a GAMLj model plots by passing new plots directives}
\usage{
gamlj_plot(gobj, haxis, sepLines = NULL, sepPlots = NULL, ...)
}
\arguments{
\item{gobj}{a gamlj results object of the class GAMLj*}

\item{haxis}{horizontal axis variable}

\item{sepLines}{variable defining the levels for separate lines}

\item{sepPlots}{variable defining the levels for which separate plots are produced}

\item{...}{any other options accepted by the gamlj_* function}
}
\value{
an object of class GAMLj* as the input object
}
\description{
This function re-estimates a GAMLj model with a new plot
}
\examples{
data<-exercise
gamljGLM(formula=)
data(qsport)
mod<-gamlj::gamljGLM(
  formula = performance ~ hours,
  data = qsport)

gamlj_plot(mod,haxis = "hours")

}
\author{
Marcello Gallucci
}
