% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{gamlj_predict}
\alias{gamlj_predict}
\title{Extract the predicted values of the model estimated by GAMLj}
\usage{
gamlj_predict(gobj)
}
\arguments{
\item{gobj}{a list of values}
}
\value{
a R object of the class of the estimated model
}
\description{
This function returns the predicted values model as a R object. It is usefull to run additional 
analysis in R with model fitted scores.
}
\examples{
data("qsport")
obj<-gamlj::gamljGLM(
   formula = performance ~ hours,
   data = qsport)
 gamlj_predict(obj)
 
}
\author{
Marcello Gallucci
}
