% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{gamlj_residuals}
\alias{gamlj_residuals}
\title{Residuals values of a GAMLj model}
\usage{
gamlj_residuals(gobj, type = "response")
}
\arguments{
\item{gobj}{a gamlj results object of the class `gamlj*Results`}

\item{type}{the type of the residuals for generalized models. The alternatives are: 'deviance' (default), 'pearson', 'working', 'response', and 'partial'. Can be abbreviated., 
cf.  \code{\link[stats:residuals.lm]{stats::residuals.lm()}}}
}
\value{
a list of value
}
\description{
Deprecated. Please use `residuals()`
}
\details{
\code{\link[stats:residuals]{stats::predict()}}, \code{\link[stats:residuals.lm]{stats::residuals.lm()}}, \code{\link[stats:residuals.glm]{stats::residuals.glm()}}
}
\examples{
data('qsport')
gmod<-gamlj::gamljGlm(
   formula = performance ~ hours,
   data = qsport)
 preds<-gamlj_residuals(gmod)
 
}
\author{
Marcello Gallucci
}
