% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{gamlj_update}
\alias{gamlj_update}
\title{Update a GAMLj model by passing only the new options}
\usage{
gamlj_update(gobj, ...)
}
\arguments{
\item{a}{single obj of class \code{\link{gamljGLM}},  \code{\link{gamljMixed}}, or  \code{\link{gamljGZLM}}}
}
\value{
an object of class GAMLj* as the input object
}
\description{
This function re-estimates a GAMLj model applying new options to the original model
}
\author{
Marcello Gallucci
}
