% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamljglmmixed.h.R
\name{gamljGlmMixed}
\alias{gamljGlmMixed}
\title{Generalized Mixed Model}
\usage{
gamljGlmMixed(
  data,
  caller = "lmer",
  modelSelection = "lmer",
  dep = NULL,
  factors = NULL,
  covs = NULL,
  modelTerms = NULL,
  fixedIntercept = TRUE,
  showParamsCI = TRUE,
  paramCIWidth = 95,
  contrasts = NULL,
  showRealNames = TRUE,
  showContrastCode = FALSE,
  plotHAxis = NULL,
  plotSepLines = NULL,
  plotSepPlots = NULL,
  plotRaw = FALSE,
  plotDvScale = FALSE,
  plotLinesTypes = FALSE,
  plotError = "none",
  ciWidth = 95,
  posthoc = NULL,
  emmeans = NULL,
  simpleVariable = NULL,
  simpleModerators = NULL,
  simpleScale = "mean_sd",
  cvalue = 1,
  percvalue = 25,
  simpleScaleLabels = "labels",
  simpleInteractions = FALSE,
  postHocCorr = list("bonf"),
  scaling = NULL,
  dep_scale = "none",
  cluster = NULL,
  randomTerms = list(list()),
  correlatedEffects = "corr",
  reml = TRUE,
  lrtRandomEffects = FALSE,
  ciRE = FALSE,
  plotRandomEffects = FALSE,
  plotOriginalScale = FALSE,
  cimethod = "wald",
  dfmethod = "Satterthwaite",
  qqplot = FALSE,
  homoTest = FALSE,
  normTest = FALSE,
  normPlot = FALSE,
  residPlot = FALSE,
  clusterBoxplot = FALSE,
  clusterResPred = FALSE,
  randHist = FALSE,
  formula
)
}
\arguments{
\item{data}{the data as a data frame}

\item{caller}{.}

\item{modelSelection}{.}

\item{dep}{a string naming the dependent variable from \code{data},
variable must be numeric}

\item{factors}{a vector of strings naming the fixed factors from
\code{data}}

\item{covs}{a vector of strings naming the covariates from \code{data}}

\item{modelTerms}{a list of character vectors describing fixed effects
terms}

\item{fixedIntercept}{\code{TRUE} (default) or \code{FALSE}, estimates
fixed intercept}

\item{showParamsCI}{\code{TRUE} (default) or \code{FALSE} , parameters CI
in table}

\item{paramCIWidth}{a number between 50 and 99.9 (default: 95) specifying
the confidence interval width for the parameter estimates}

\item{contrasts}{a list of lists specifying the factor and type of contrast
to use, one of \code{'deviation'}, \code{'simple'}, \code{'difference'},
\code{'helmert'}, \code{'repeated'} or \code{'polynomial'}}

\item{showRealNames}{\code{TRUE} or \code{FALSE} (default), provide raw
names of the contrasts variables}

\item{showContrastCode}{\code{TRUE} or \code{FALSE} (default), provide
contrast coefficients tables}

\item{plotHAxis}{a string naming the variable placed on the horizontal axis
of the plot}

\item{plotSepLines}{a string naming the variable represented as separate
lines in the plot}

\item{plotSepPlots}{a list of string naming the variables defining the
levels for multiple plots}

\item{plotRaw}{\code{TRUE} or \code{FALSE} (default), provide descriptive
statistics}

\item{plotDvScale}{\code{TRUE} or \code{FALSE} (default), scale the plot
Y-Axis to the max and the min of the dependent variable observed scores.}

\item{plotLinesTypes}{\code{TRUE} or \code{FALSE} (default), use different
linetypes per levels.}

\item{plotError}{\code{'none'}, \code{'ci'} (default), or \code{'se'}. Use
no error bars, use confidence intervals, or use standard errors on the
plots, respectively}

\item{ciWidth}{a number between 50 and 99.9 (default: 95) specifying the
confidence interval width}

\item{posthoc}{a rhs formula with the terms specifying the table to apply
the comparisons (of the form \code{'~x+x:z'})}

\item{emmeans}{a rhs formula with the terms specifying the marginal means
to estimate (of the form \code{'~x+x:z'})}

\item{simpleVariable}{The variable for which the simple effects (slopes)
are computed}

\item{simpleModerators}{the variable that provides the levels at which the
simple effects computed}

\item{simpleScale}{\code{'mean_sd'} (default), \code{'custom'} , or
\code{'custom_percent'}. Use to condition the covariates (if any)}

\item{cvalue}{offset value for conditioning}

\item{percvalue}{offset value for conditioning}

\item{simpleScaleLabels}{decide the labeling of simple effects in tables
and plots.  \code{labels} indicates that only labels are used, such as
\code{Mean} and  \code{Mean + 1 SD}. \code{values} uses the actual values
as labels. \code{values_labels} uses both.}

\item{simpleInteractions}{should simple Interactions be computed}

\item{postHocCorr}{one or more of \code{'none'},  \code{'bonf'}, or
\code{'holm'}; provide no,  Bonferroni, and Holm Post Hoc corrections
respectively}

\item{scaling}{a list of lists specifying the covariates scaling, one of
\code{'centered to the mean'}, \code{'standardized'}, or \code{'none'}.
\code{'none'} leaves the variable as it is}

\item{dep_scale}{Re-scale the dependent variable.}

\item{cluster}{a vector of strings naming the clustering variables from
\code{data}}

\item{randomTerms}{a list of lists specifying the models random effects.}

\item{correlatedEffects}{\code{'nocorr'}, \code{'corr'} (default), or
\code{'block'}. When random effects are passed as list of length 1, it
decides whether the effects should be correlated,  non correlated. If
\code{'randomTerms'} is a list of  lists of length > 1, the option is
automatially set to \code{'block'}. The option is ignored if the model is
passed using \code{formula}.}

\item{reml}{\code{TRUE} (default) or \code{FALSE}, should the Restricted ML
be used rather than ML}

\item{lrtRandomEffects}{\code{TRUE} or \code{FALSE} (default), LRT for the
random effects}

\item{ciRE}{\code{TRUE} or \code{FALSE} (default), confidence intervals for
the random effects}

\item{plotRandomEffects}{\code{TRUE} or \code{FALSE} (default), add random
effects predicted values in the plot}

\item{plotOriginalScale}{\code{TRUE} or \code{FALSE} (default), use
original scale for covariates.}

\item{cimethod}{.}

\item{dfmethod}{.}

\item{qqplot}{\code{TRUE} or \code{FALSE} (default), provide a Q-Q plot of
residuals}

\item{homoTest}{\code{TRUE} or \code{FALSE} (default), perform homogeneity
tests}

\item{normTest}{\code{TRUE} or \code{FALSE} (default), provide a test for
normality of residuals}

\item{normPlot}{\code{TRUE} or \code{FALSE} (default), provide a histogram
of residuals superimposed by a normal distribution}

\item{residPlot}{\code{TRUE} or \code{FALSE} (default), provide a
scatterplot of the residuals against predicted}

\item{clusterBoxplot}{\code{TRUE} or \code{FALSE} (default), provide a
boxplot of random effects by the first defined clustering variable}

\item{clusterResPred}{\code{TRUE} or \code{FALSE} (default), residuals vs
predicted by the first defined clustering variable}

\item{randHist}{\code{TRUE} or \code{FALSE} (default), provide histogram of
random Coefficients}

\item{formula}{(optional) the formula to use, see the examples}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$model} \tab \tab \tab \tab \tab The underlying \code{lmer} object \cr
  \code{results$info} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$r2} \tab \tab \tab \tab \tab a table of R \cr
  \code{results$main$fit} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$anova} \tab \tab \tab \tab \tab a table of ANOVA results \cr
  \code{results$main$coefficients} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$random} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$randomCov} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$randomTests} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$contrastCodeTables} \tab \tab \tab \tab \tab an array of contrast coefficients tables \cr
  \code{results$posthoc} \tab \tab \tab \tab \tab an array of post-hoc tables \cr
  \code{results$simpleEffects$anova} \tab \tab \tab \tab \tab a table of ANOVA for simple effects \cr
  \code{results$simpleEffects$coefficients} \tab \tab \tab \tab \tab a table \cr
  \code{results$simpleInteractions} \tab \tab \tab \tab \tab an array of simple interactions tables \cr
  \code{results$emmeans} \tab \tab \tab \tab \tab an array of predicted means tables \cr
  \code{results$mainPlots} \tab \tab \tab \tab \tab an array of results plots \cr
  \code{results$plotnotes} \tab \tab \tab \tab \tab a html \cr
  \code{results$assumptions$homoTest} \tab \tab \tab \tab \tab a table of homogeneity tests \cr
  \code{results$assumptions$normTest} \tab \tab \tab \tab \tab a table of normality tests \cr
  \code{results$assumptions$qqplot} \tab \tab \tab \tab \tab a q-q plot \cr
  \code{results$assumptions$normPlot} \tab \tab \tab \tab \tab Residual histogram \cr
  \code{results$assumptions$residPlot} \tab \tab \tab \tab \tab Residual Predicted plot \cr
  \code{results$assumptions$clusterResPred} \tab \tab \tab \tab \tab Residuals vs Predicted by cluster \cr
  \code{results$assumptions$clusterBoxplot} \tab \tab \tab \tab \tab Residuals boxplot by cluster \cr
  \code{results$assumptions$randHist} \tab \tab \tab \tab \tab an array of random coefficients histograms \cr
  \code{results$predicted} \tab \tab \tab \tab \tab an output \cr
  \code{results$residuals} \tab \tab \tab \tab \tab an output \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$info$asDF}

\code{as.data.frame(results$info)}
}
\description{
Mixed Linear Model
}
\examples{
data(subjects_by_stimuli)

gamlj::gamljMixed(
       formula = y ~ 1 + cond+( 1|subj ),
       data = subjects_by_stimuli)

gamlj::gamljMixed(
       data = subjects_by_stimuli,
       dep = "y",
       factors = "cond",
       modelTerms = "cond",
       cluster = "subj",
       randomTerms=list(list(c("cond","subj"))))

}
