% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{get_data}
\alias{get_data}
\alias{get_data.gamlj}
\title{Extract data}
\usage{
get_data(object, ...)
}
\arguments{
\item{object}{a gamlj results object of the class `gamlj*Results`}

\item{...}{additional arguments passed to the GAMLj estimation function}
}
\value{
a dataset
}
\description{
This function returns a dataset with the variables in the GAMLj model 
transformed according to GAMLj options. It is usefull to run additional 
models with other R packages with the same setup used by GAMLj
}
\examples{
data('qsport')
gmod<-GAMLj3::gamlj_lm(formula = performance ~ hours,
                data = qsport,
                covs_scale = c(hours='standardized'))

gdata<-get_data(gmod)
lm(performance ~ hours,data=gdata)
}
\author{
Marcello Gallucci
}
