% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gmixeddata}
\alias{gmixeddata}
\title{Clustered data for generalized mixed model}
\usage{
data(glmmixeddata)
}
\description{
Generated data to test generalized mixed models. The variables can be used:
 `x` and `z` as two continuous independent variables; `w` as categorical independent variable;
`ybin` as dichotomous dependent variable; `ypoi` as count variable (Poisson);
`cluster` as the clustering variable.
}
\examples{
data(glmmixeddata)
}
\keyword{datasets}
