% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{manymodels}
\alias{manymodels}
\title{Many Models Data}
\usage{
data(manymodels)
}
\description{
Generated data to test linear models. The variables can be used as:
}
\details{
`x` and `z` as two continuous independent variables;

      `cat2` and `cat3`  as categorical independent variables, with two and three groups respectively;

      `ycont` is a continuous variable, suitable as dependent variable.
      `ybin` is dichotomous dependent variable; `ypoi` a count variable (Poisson),
      `yord` and ordinal dependent variable with 5 levels, and `ycat` a categorical dependent variable with three groups.
}
\examples{
data(manymodels)
}
\keyword{datasets}
