% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phdpubs}
\alias{phdpubs}
\title{Clustered data for Negative Binomial model}
\usage{
data(phdpubs)
}
\description{
Modified version of a dataset in Long, J. S., & Freese, J. (2006). Regression models for categorical dependent variables using Stata. Stata press.
Data present number of articles published by PhD students in a three year period with some predictors. Data are clustered by program.
}
\examples{
data(phdpubs)
}
\keyword{datasets}
