% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poissonacts}
\alias{poissonacts}
\title{Poisson data}
\usage{
data(poissonacts)
}
\description{
Simulated data of a Poisson distributed dependent variables with some independent variabes
}
\examples{
data(poissonacts)
}
\keyword{datasets}
