% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qsport}
\alias{qsport}
\title{Sport data with curvilinear effects}
\usage{
data(qsport)
}
\description{
Simulated data for curvilinear effects examples
}
\examples{
data(qsport)
}
\keyword{datasets}
