% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer.rsquared.R
\name{r.squared.lm}
\alias{r.squared.lm}
\title{Marginal r-squared for lm objects}
\usage{
\method{r.squared}{lm}(mdl)
}
\arguments{
\item{mdl}{an lm object (usually fit using \code{\link{lm}},}
}
\value{
a dataframe with with "Class" = "lm", "Family" = "gaussian",
       "Marginal" = unadjusted r-squared, "Conditional" = NA, and "AIC" columns
}
\description{
This method uses r.squared from \code{\link{summary}} as the marginal.
Contrary to other \code{\link{r.squared}} methods, 
this one doesn't call \code{\link{.rsquared.glmm}}
}
