% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer.rsquared.R
\name{r.squared.lme}
\alias{r.squared.lme}
\title{Marginal and conditional r-squared for lme objects}
\usage{
\method{r.squared}{lme}(mdl)
}
\arguments{
\item{mdl}{an lme model (usually fit using \code{\link{nlme::lme}})}
}
\description{
This method extracts the variance for fixed and random effects,
as well as residuals, and calls \code{\link{.rsquared.glmm}}
}
