% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer.rsquared.R
\name{r.squared.merMod}
\alias{r.squared.merMod}
\title{Marginal and conditional r-squared for merMod objects}
\usage{
\method{r.squared}{merMod}(mdl)
}
\arguments{
\item{mdl}{an merMod model (usually fit using \code{\link{lme4::lmer}},
\code{\link{lme4::glmer}}, \code{\link{lmerTest::lmer}},
\code{\link{blme::blmer}}, \code{\link{blme::bglmer}}, etc)}
}
\description{
This method extracts the variance for fixed and random effects, residuals,
and the fixed effects for the null model (in the case of Poisson family),
and calls \code{\link{.rsquared.glmm}}
}
