% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer.rsquared.R
\name{r.squared}
\alias{r.squared}
\title{R-squared and pseudo-rsquared for (generalized) linear (mixed) models}
\usage{
r.squared(mdl)
}
\arguments{
\item{mdl}{a fitted (generalized) linear (mixed) model object}
}
\value{
Implementing methods usually return a dataframe with "Class",
        "Family", "Marginal", "Conditional", and "AIC" columns
}
\description{
This generic function calculates the r squared and pseudo r-squared for
a variety of(generalized) linear (mixed) model fits.
Currently implemented for \code{\link{lm}}, \code{\link{lmerTest::merMod}},
and \code{\link{nlme::lme}} objects.
Implementing methods usually call \code{\link{.rsquared.glmm}}
}
