% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{gamlj_residuals}
\alias{gamlj_residuals}
\alias{residuals}
\alias{residuals.gamljGlmResults}
\alias{residuals.gamljGzlmResults}
\alias{residuals.gamljMixedResults}
\alias{residuals.gamljGlmMixedResults}
\title{Residuals values of a GAMLj model}
\usage{
gamlj_residuals(gobj, type = "response")

\method{residuals}{gamljGlmResults}(object, ...)

\method{residuals}{gamljGzlmResults}(object, type = "deviance", ...)

\method{residuals}{gamljMixedResults}(object, ...)

\method{residuals}{gamljGlmMixedResults}(object, type = "deviance", ...)
}
\arguments{
\item{gobj}{a gamlj results object of the class `gamlj*Results`}

\item{type}{the type of residuals for generalized models. The alternatives are: "deviance" (default), "pearson", "working", "response", and "partial". Can be abbreviated.
Cf. \code{\link[stats:residuals]{stats::residuals()}}, \code{\link[stats:residuals.lm]{stats::residuals.lm()}}, \code{\link[stats:residuals.glm]{stats::residuals.glm()}}}

\item{object}{a gamlj results object of the class `gamlj*Results`}

\item{...}{additional arguments for specific residuals methods.}
}
\value{
a list of value

a R object of the class of the estimated model
}
\description{
Deprecated. Please use `residuals()`

Returns residuals values from the estimated model
}
\details{
\code{\link[stats:residuals]{stats::predict()}}, \code{\link[stats:residuals.lm]{stats::residuals.lm()}}, \code{\link[stats:residuals.glm]{stats::residuals.glm()}}
}
\examples{
data("qsport")
obj<-gamlj::gamljGlm(
   formula = performance ~ hours,
   data = qsport)
 preds<-gamlj_residuals(obj)
 
data("qsport")
obj<-gamlj::gamljGlm(
   formula = performance ~ hours,
   data = qsport)
 preds<-residuals(obj)
 
}
\author{
Marcello Gallucci

Marcello Gallucci
}
