% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer.rsquared.R
\name{rsquared.glmm}
\alias{rsquared.glmm}
\title{R-squared and pseudo-rsquared for a list of (generalized) linear (mixed) models}
\usage{
rsquared.glmm(modlist)
}
\arguments{
\item{a}{single model or a list of fitted (generalized) linear (mixed) model objects}
}
\value{
a dataframe with one row per model, and "Class",
        "Family", "Marginal", "Conditional" and "AIC" columns
}
\description{
This function calls the generic \code{\link{r.squared}} function for each of the
models in the list and rbinds the outputs into one data frame
}
