% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{summary.ResultsElement}
\alias{summary.ResultsElement}
\alias{print.jmvrtable}
\alias{print.jmvrobj}
\alias{coef.gamlj}
\alias{fit}
\alias{fit.gamlj}
\title{S3 methods for class jamovi ResultsElement}
\usage{
\method{summary}{ResultsElement}(object, ...)

\method{print}{jmvrtable}(x, ...)

\method{print}{jmvrobj}(x, ...)

\method{coef}{gamlj}(x, ...)

fit(x, ...)

\method{fit}{gamlj}(x, ...)
}
\arguments{
\item{object}{a gamlj results object of the class `gamlj`}

\item{x}{a gamlj results object of the class `gamlj`}
}
\value{
a list of table as data.frame
}
\description{
These functions extract all visible tables from a ResultsElement produced by gamlj
and print them in R style.
}
\examples{
data(fivegroups)
fivegroups$Group<-factor(fivegroups$Group)
gmod<-gamlj::gamljGlm(
  formula = Score ~Group,
  data = fivegroups)

summary(gmod)
}
\author{
Marcello Gallucci
}
