% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schoolexam}
\alias{schoolexam}
\title{Clustered data for logistic regression}
\usage{
data(schoolexam)
}
\description{
Simulated data for testing logistic mixed models. 
Simulation of pupils data clustered across schools with a dichotomous outcome
}
\examples{
data(schoolexam)
}
\keyword{datasets}
