% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{simpleeffects}
\alias{simpleeffects}
\alias{simpleeffects.gamlj}
\title{Simple Effects on GAMLj results}
\usage{
simpleeffects(object, ...)

\method{simpleeffects}{gamlj}(object, simple_effects = NULL, simple_moderators = NULL, ...)
}
\arguments{
\item{object}{a gamlj results object of the class `gamlj`}

\item{...}{all options accepted by a gamlj model function. Relevant for new tests are 
`simpleVariable` (the simple effect variable), `simpleModerator` (the moderator), and `simple3way` for the second moderator.}

\item{formula}{a right hand side formula specifying the variables to test, of the form `~x:z`, `~x:z:w` or `~x*z`. 
The formula is not expanded, so the first variable is the simple effect variable, the second is the moderator, 
the third an optional additional moderator, an so on
It has prevalence on other options defining a simple effects test via character options.}
}
\value{
an object of class gamlj results
}
\description{
This is a convenience function to re-estimates a GAMLj model adding simple effect analysis. If no options is passed, extracts the 
simple effects tables already in the model results (if any). If new tests are defined, the simple effects tests tables 
are returned.
}
\examples{
data(winkel)
wicksell$time<-factor(wicksell$time)
wicksell$group<-factor(wicksell$group) 
gmod<-gamlj::gamljMixed(
   formula = dv ~ 1 +group+ time:group+ time+( 1 | subj ),
   data = wicksell)

simpleEffects(gmod,formula =~time:group)
}
\author{
Marcello Gallucci
}
