% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subjects_by_stimuli}
\alias{subjects_by_stimuli}
\title{Subjects by stimuli cross-classification dataset}
\usage{
data(subjects_by_stimuli)
}
\description{
Simulated data for subjects by stimuli cross-classification examples
}
\examples{
data(subjects_by_stimuli)
}
\keyword{datasets}
