% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subjects_on_stimuli}
\alias{subjects_on_stimuli}
\title{Subjects by stimuli nested classification dataset}
\usage{
data(subjects_on_stimuli)
}
\description{
Data for stimuli nested in subjects classification examples
}
\examples{
data(subjects_on_stimuli)
}
\keyword{datasets}
