% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temptime}
\alias{temptime}
\title{Individual growth model data}
\usage{
data(temptime)

data(temptime)
}
\description{
Data for Individual growth model data examples

Simulated data for individual growth models. 
Simulation of an experiment in which two procedures are administered 
between-subjects to two different groups. Each participant’s body temperature
is measured for about 6 minutes every 6 seconds. Time is recorded in tenth of minutes
(6 seconds=0.1, 12 seconds=0.2, etc.).
}
\examples{
data(temptime)
data(temptime)
}
\keyword{datasets}
