% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{test_contrasts}
\alias{test_contrasts}
\title{Custom contrasts}
\usage{
test_contrasts(object, contrasts, ...)
}
\arguments{
\item{object}{a gamlj results object of the class `gamlj`}

\item{contrasts}{a named list with contrasts weights to test of the form \code{list(factor1=c(a1,a2,b3), factor2=c(b1,b2,b3))}. Multiple
contrasts for the same factor can be tested. In this case, each contrast is tested independently.}

\item{...}{all options accepted by a gamlj model function.}
}
\value{
an object of class
}
\description{
This is a convenience function to test and extract estimates and information about
custom contrast.
}
\examples{
data(fivegroups)
fivegroups$Group<-factor(fivegroups$Group)
gmod<-GAMLj3::gamlj_lm(
  formula = Score ~Group,
  data = fivegroups)

em_means(gmod,formula =~Group)
}
\author{
Marcello Gallucci
}
