% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{update.gamljlmResults}
\alias{update.gamljlmResults}
\alias{update.gamljglmResults}
\alias{update.gamljmixedResults}
\alias{update.gamljgmixedResults}
\title{Update a GAMLj model}
\usage{
\method{update}{gamljlmResults}(object, ...)

\method{update}{gamljglmResults}(object, ...)

\method{update}{gamljmixedResults}(object, ...)

\method{update}{gamljgmixedResults}(object, ...)
}
\arguments{
\item{object}{of class `gamlj*Results`}

\item{...}{any parameter to be passed to \code{\link{gamljGlm}},  \code{\link{gamljmixed}},  \code{\link{gamljglm}}, or  \code{\link{gamljgmixed}}}
}
\value{
an object of class gamlj*Results as the input object
}
\description{
Re-estimates a GAMLj model applying new options to the original model
}
\author{
Marcello Gallucci
}
