% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wicksell}
\alias{wicksell}
\title{Depression over time}
\usage{
data(wicksell)
}
\description{
Data repeated measure anova with mixed models
}
\examples{
data(wicksell)
}
\references{
David C. Howell, Overview of Mixed Models \url{https://www.uvm.edu/~statdhtx/StatPages/Mixed-Models-Repeated/Mixed-Models-Overview.html}
}
\keyword{datasets}
