% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{gamlj_ggplot}
\alias{gamlj_ggplot}
\title{This function returns the plot as a ggplot object}
\usage{
gamlj_ggplot(gobj)
}
\arguments{
\item{gobj}{a gamlj results object of the class GAMLj*#'}
}
\value{
an object of ggplot
}
\description{
This function returns the plot as a ggplot object
}
\author{
Marcello Gallucci
}
