% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{gamlj_ggplot}
\alias{gamlj_ggplot}
\title{Extract a ggplot2 plot from a GAMLj results object
 
This function returns the plot as a ggplot object contained in the gamlj object (the model). If the model 
contains one plot, it is returned. If there are more plots, a list is returned with plots as elements of the list.}
\usage{
gamlj_ggplot(gobj)
}
\arguments{
\item{gobj}{a gamlj results object of the class GAMLj*#'}
}
\value{
an object of ggplot
}
\description{
Extract a ggplot2 plot from a GAMLj results object
 
This function returns the plot as a ggplot object contained in the gamlj object (the model). If the model 
contains one plot, it is returned. If there are more plots, a list is returned with plots as elements of the list.
}
\author{
Marcello Gallucci
}
