% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{gamlj_ggplot}
\alias{gamlj_ggplot}
\title{This function returns the plot as a ggplot object}
\usage{
gamlj_ggplot(gobj, ...)
}
\arguments{
\item{gobj}{a gamlj results object of the class GAMLj*#'}

\item{...}{any other options accepted by the gamlj_* function}

\item{theme}{a theme for the ggplot plot. Default is the jamovi default theme}
}
\value{
an object of class GAMLj* as the input object
}
\description{
This function returns the plot as a ggplot object
}
\author{
Marcello Gallucci
}
