% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{gamlj_model}
\alias{gamlj_model}
\title{Extract the model estimated by GAMLj}
\usage{
gamlj_model(gobj)
}
\arguments{
\item{gobj}{a gamlj results object of the class GAMLj*}
}
\value{
a R object of the class of the estimated model
}
\description{
This function returns the estimated model as a R object. It is usefull to run additional 
analysis in R with model object.
}
\examples{
data("qsport")
obj<-gamlj::gamljGLM(
   formula = performance ~ hours,
   data = qsport)
}
\author{
Marcello Gallucci
}
