% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{gamlj_model}
\alias{gamlj_model}
\title{Extract the R model object contained in the GAMLj results}
\usage{
gamlj_model(object)
}
\arguments{
\item{object}{a gamlj results object of the class `gamlj*Results`}
}
\value{
a R object from lm(), glm(), lmer() or glmer()
}
\description{
Return the estimated model object of the class lm, lmer, glm depending on the model.
}
\examples{
data('qsport')
obj<-gamlj::gamljGlm(
   formula = performance ~ hours,
   data = qsport)
 model<-gamlj_model(obj)
 
}
\author{
Marcello Gallucci
}
