% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{gamlj_plot}
\alias{gamlj_plot}
\title{Update a GAMLj model plots by passing new plots directives}
\usage{
gamlj_plot(gobj, haxis, seplines = NULL, sepplots = NULL)
}
\arguments{
\item{haxis}{horizontal axis variable}

\item{seplines}{variable defining the levels for separate lines}

\item{sepplot}{variable defining the levels for which separate plots are produced}
}
\value{
an object of class GAMLj* as the input object
}
\description{
This function re-estimates a GAMLj model with a new plot
}
\author{
Marcello Gallucci
}
