% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{gamlj_residuals}
\alias{gamlj_residuals}
\title{Residuals values of a GAMLj model}
\usage{
gamlj_residuals(gobj, type = "response")
}
\arguments{
\item{gobj}{a gamlj results object of the class GAMLj*}

\item{type}{the type of the residuals. cf.  it can be "response" (default),"working", "deviance", "pearson",
"partial", cf.  \code{\link[stats:residuals.lm]{stats::residuals.lm()}}}
}
\value{
a list of value
}
\description{
This function returns the residuals of a GAMLj model. It is usefull to run additional 
analysis on model assumptions. 
\code{\link[stats:residuals]{stats::predict()}}, \code{\link[stats:residuals.lm]{stats::residuals.lm()}}, \code{\link[stats:residuals.glm]{stats::residuals.glm()}}
}
\examples{
data("qsport")
obj<-gamlj::gamljGlm(
   formula = performance ~ hours,
   data = qsport)
 preds<-gamlj_residuals(obj)
 
}
\author{
Marcello Gallucci
}
