% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{gamlj_residuals}
\alias{gamlj_residuals}
\title{Extract the residuals values of the model estimated by GAMLj}
\usage{
gamlj_residuals(gobj)
}
\arguments{
\item{gobj}{a gamlj results object of the class GAMLj*}

\item{type}{the type of the residuals. cf. [stats::residuals()]}
}
\value{
a list of value
}
\description{
This function returns the residuals of a model as a R object. It is usefull to run additional 
analysis in R on model assumptions. 
\code{\link[stats:residuals]{stats::predict()}}, \code{\link[stats:residuals.lm]{stats::residuals.lm()}}, \code{\link[stats:residuals.glm]{stats::residuals.glm()}}
}
\examples{
data("qsport")
obj<-gamlj::gamljGLM(
   formula = performance ~ hours,
   data = qsport)
 preds<-gamlj_residuals(obj)
 
}
\author{
Marcello Gallucci
}
