% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamljglm.h.R
\name{gamljGLM}
\alias{gamljGLM}
\title{General Linear Model}
\usage{
gamljGLM(data, dep = NULL, factors = NULL, covs = NULL,
  modelTerms = NULL, ss = "3", effectSize = NULL, showParamsCI = TRUE,
  paramCIWidth = 95, contrasts = NULL, showContrastsTable = FALSE,
  showContrasts = TRUE, scaling = NULL, plotHAxis = NULL,
  plotSepLines = NULL, plotSepPlots = NULL, plotRaw = FALSE,
  plotDvScale = FALSE, postHoc = NULL, postHocCorr = list("tukey"),
  eDesc = FALSE, homo = FALSE, qq = FALSE, plotError = "ci",
  ciWidth = 95, simpleVariable = NULL, simpleModerator = NULL,
  simple3way = NULL)
}
\arguments{
\item{data}{the data as a data frame}

\item{dep}{a string naming the dependent variable from \code{data}, 
variable must be numeric}

\item{factors}{a vector of strings naming the fixed factors from 
\code{data}}

\item{covs}{a vector of strings naming the covariates from \code{data}}

\item{modelTerms}{a list of character vectors describing the terms to go 
into the model}

\item{ss}{\code{'1'}, \code{'2'} or \code{'3'} (default), the sum of 
squares to use}

\item{effectSize}{one or more of \code{'eta'}, \code{'partEta'}, or 
\code{'omega'}; use eta², partial eta², and omega² effect sizes, 
respectively}

\item{showParamsCI}{\code{TRUE} or \code{FALSE} (default), parameters CI in 
table}

\item{paramCIWidth}{a number between 50 and 99.9 (default: 95) specifying 
the confidence interval width for the parameter estimates}

\item{contrasts}{a list of lists specifying the factor and type of contrast 
to use, one of \code{'deviation'}, \code{'simple'}, \code{'difference'}, 
\code{'helmert'}, \code{'repeated'} or \code{'polynomial'}}

\item{showContrastsTable}{\code{TRUE} or \code{FALSE} (default), provide 
definitions of the contrasts variables}

\item{showContrasts}{\code{TRUE} or \code{FALSE} (default), provide 
definitions of the contrasts variables}

\item{scaling}{a list of lists specifying the covariates scaling, one of 
\code{'centered to the mean'}, \code{'standardized'}, or \code{'none'}. 
\code{'none'} leaves the variable as it is}

\item{plotHAxis}{a string naming the variable placed on the horizontal axis 
of the plot}

\item{plotSepLines}{a string naming the variable represented as separate 
lines on the plot}

\item{plotSepPlots}{a string naming the variable to separate over to form 
multiple plots}

\item{plotRaw}{.}

\item{plotDvScale}{.}

\item{postHoc}{a list of terms to perform post-hoc tests on}

\item{postHocCorr}{one or more of \code{'none'}, \code{'tukey'}, 
\code{'scheffe'}, \code{'bonf'}, or \code{'holm'}; provide no, Tukey, 
Scheffe, Bonferroni, and Holm Post Hoc corrections respectively}

\item{eDesc}{\code{TRUE} or \code{FALSE} (default), provide descriptive 
statistics}

\item{homo}{\code{TRUE} or \code{FALSE} (default), perform homogeneity 
tests}

\item{qq}{\code{TRUE} or \code{FALSE} (default), provide a Q-Q plot of 
residuals}

\item{plotError}{\code{'none'}, \code{'ci'} (default), or \code{'se'}. Use 
no error bars, use confidence intervals, or use standard errors on the 
plots, respectively}

\item{ciWidth}{a number between 50 and 99.9 (default: 95) specifying the 
confidence interval width}

\item{simpleVariable}{The variable for which the simple effects (slopes) 
are computed}

\item{simpleModerator}{the variable that provides the levels at which the 
simple effects computed}

\item{simple3way}{a moderator of the two-way interaction which is probed}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$main} \tab \tab \tab \tab \tab a table of ANOVA results \cr
  \code{results$estimates} \tab \tab \tab \tab \tab a table of Coefficients \cr
  \code{results$contrasts} \tab \tab \tab \tab \tab an array of contrasts definitions tables \cr
  \code{results$simpleEffectsAnovas} \tab \tab \tab \tab \tab an array of simple Effects ANOVA tables \cr
  \code{results$simpleEffects} \tab \tab \tab \tab \tab an array of simple Effects tables \cr
  \code{results$model} \tab \tab \tab \tab \tab The underlying \code{lm} object \cr
  \code{results$assump$homo} \tab \tab \tab \tab \tab a table of homogeneity tests \cr
  \code{results$assump$qq} \tab \tab \tab \tab \tab a q-q plot \cr
  \code{results$postHoc} \tab \tab \tab \tab \tab an array of post-hoc tables \cr
  \code{results$emeansTables} \tab \tab \tab \tab \tab an array of predicted means tables \cr
  \code{results$descPlot} \tab \tab \tab \tab \tab a descriptives plot \cr
  \code{results$descPlots} \tab \tab \tab \tab \tab an array of results plots \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$main$asDF}

\code{as.data.frame(results$main)}
}
\description{
General Linear Model
}
\examples{
data('data3by2')

gamljGLM(dat, dep = 'y', factors =c('twogroups'), covs = 'x')

#
#  
#
#  ANOVA
#  -----------------------------------------------------------------------
#                 Sum of Squares    df    Mean Square    F        p
#  -----------------------------------------------------------------------
#    x               115.19         1      115.19       20.07     < .001
#    twogroups       31.36          1      31.36         5.46      0.021  
#    x:twogroups     7.15          1       7.15         1.25      0.267  
#    Residuals      551.13         96       5.74  
#  -----------------------------------------------------------------------
#
#  _______________________________________________________________________
#                  Estimates  Std. Error  t-value      p.
#  _______________________________________________________________________
#
#  (Intercept)     52.804      0.245        215.17   < .001
#   x              0.596       0.133        4.48     < .001
#   twogroups1    -0.574       0.245       -2.34      0.021
#   x:twogroups1  -0.148       0.133       -1.12      0.267

}
