% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamljglmmixed.h.R
\name{gamljGlmMixed}
\alias{gamljGlmMixed}
\title{Generalized Mixed Models}
\usage{
gamljGlmMixed(
  data,
  dep = NULL,
  factors = NULL,
  covs = NULL,
  modelTerms = NULL,
  fixedIntercept = TRUE,
  showParamsCI = FALSE,
  showExpbCI = TRUE,
  paramCIWidth = 95,
  contrasts = NULL,
  showRealNames = TRUE,
  showContrastCode = FALSE,
  plotHAxis = NULL,
  plotSepLines = NULL,
  plotSepPlots = NULL,
  plotRaw = FALSE,
  plotDvScale = FALSE,
  plotError = "none",
  ciWidth = 95,
  postHoc = NULL,
  eDesc = FALSE,
  eCovs = FALSE,
  simpleVariable = NULL,
  simpleModerator = NULL,
  simple3way = NULL,
  simpleScale = "mean_sd",
  cvalue = 1,
  percvalue = 25,
  simpleScaleLabels = "labels",
  postHocCorr = list("bonf"),
  scaling = NULL,
  cluster = NULL,
  randomTerms = list(list()),
  correlatedEffects = "corr",
  plotRandomEffects = FALSE,
  plotLinearPred = FALSE,
  nAGQ = 1,
  effectSize = list("expb"),
  modelSelection = "logistic",
  custom_family = "gaussian",
  custom_link = "identity",
  cimethod,
  formula
)
}
\arguments{
\item{data}{the data as a data frame}

\item{dep}{a string naming the dependent variable from \code{data},
variable must be numeric}

\item{factors}{a vector of strings naming the fixed factors from
\code{data}}

\item{covs}{a vector of strings naming the covariates from \code{data}}

\item{modelTerms}{a list of character vectors describing fixed effects
terms}

\item{fixedIntercept}{\code{TRUE} (default) or \code{FALSE}, estimates
fixed intercept}

\item{showParamsCI}{\code{TRUE} (default) or \code{FALSE} , parameters CI
in table}

\item{showExpbCI}{\code{TRUE} (default) or \code{FALSE} , exp(B) CI in
table}

\item{paramCIWidth}{a number between 50 and 99.9 (default: 95) specifying
the confidence interval width for the parameter estimates}

\item{contrasts}{a list of lists specifying the factor and type of contrast
to use, one of \code{'deviation'}, \code{'simple'}, \code{'difference'},
\code{'helmert'}, \code{'repeated'} or \code{'polynomial'}}

\item{showRealNames}{\code{TRUE} or \code{FALSE} (default), provide raw
names of the contrasts variables}

\item{showContrastCode}{\code{TRUE} or \code{FALSE} (default), provide
contrast coefficients tables}

\item{plotHAxis}{a string naming the variable placed on the horizontal axis
of the plot}

\item{plotSepLines}{a string naming the variable represented as separate
lines on the plot}

\item{plotSepPlots}{the variable for whose levels multiple plots are
computed}

\item{plotRaw}{\code{TRUE} or \code{FALSE} (default), provide descriptive
statistics}

\item{plotDvScale}{\code{TRUE} or \code{FALSE} (default), scale the plot
Y-Axis to the max and the min of the dependent variable observed scores.}

\item{plotError}{\code{'none'}, \code{'ci'} (default), or \code{'se'}. Use
no error bars, use confidence intervals, or use standard errors on the
plots, respectively}

\item{ciWidth}{a number between 50 and 99.9 (default: 95) specifying the
confidence interval width}

\item{postHoc}{a list of terms to perform post-hoc tests on}

\item{eDesc}{\code{TRUE} or \code{FALSE} (default), provide lsmeans
statistics}

\item{eCovs}{\code{TRUE} or \code{FALSE} (default), provide lsmeans
statistics}

\item{simpleVariable}{The variable for which the simple effects (slopes)
are computed}

\item{simpleModerator}{the variable that provides the levels at which the
simple effects computed}

\item{simple3way}{a moderator of the two-way interaction which is probed}

\item{simpleScale}{\code{'mean_sd'} (default), \code{'custom'} , or
\code{'custom_percent'}. Use to condition the covariates (if any)}

\item{cvalue}{offset value for conditioning}

\item{percvalue}{offset value for conditioning}

\item{simpleScaleLabels}{decide the labeling of simple effects in tables
and plots.  \code{labels} indicates that only labels are used, such as
\code{Mean} and  \code{Mean + 1 SD}. \code{values} uses the actual values
as labels. \code{values_labels} uses both.}

\item{postHocCorr}{one or more of \code{'none'},  \code{'bonf'}, or
\code{'holm'}; provide no,  Bonferroni, and Holm Post Hoc corrections
respectively}

\item{scaling}{a list of lists specifying the covariates scaling, one of
\code{'centered to the mean'}, \code{'standardized'}, or \code{'none'}.
\code{'none'} leaves the variable as it is}

\item{cluster}{a vector of strings naming the clustering variables from
\code{data}}

\item{randomTerms}{a list of lists specifying the models random effects.}

\item{correlatedEffects}{\code{'nocorr'}, \code{'corr'} (default), or
\code{'block'}. When random effects are passed as list of length 1, it
decides whether the effects should be correlated,  non correlated. If
\code{'randomTerms'} is a list of  lists of length > 1, the option is
automatially set to \code{'block'}. The option is ignored if the model is
passed using \code{formula}.}

\item{plotRandomEffects}{\code{TRUE} or \code{FALSE} (default), add random
effects predicted values in the plot}

\item{plotLinearPred}{\code{TRUE} or \code{FALSE} (default), plot the
predicted values in the linear predictor scale. If set, \code{plotRaw} and
\code{plotDvScale} are ignored.}

\item{nAGQ}{for the adaptive Gauss-Hermite log-likelihood approximation,
nAGQ argument controls the number of nodes in the quadrature formula. A
positive integer}

\item{effectSize}{.}

\item{modelSelection}{Select the generalized linear model:
\code{linear},\code{poisson},\code{logistic},\code{multinomial}}

\item{custom_family}{Distribution family for the custom model, accepts
gaussian, binomial, gamma and inverse_gaussian .}

\item{custom_link}{Distribution family for the custom model, accepts
identity, log and inverse, onemu2 (for 1/mu^2).}

\item{cimethod}{.}

\item{formula}{(optional) the formula to use, see the examples}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$info} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$anova} \tab \tab \tab \tab \tab a table of ANOVA results \cr
  \code{results$main$fixed} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$random} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$randomCov} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$contrastCodeTables} \tab \tab \tab \tab \tab an array of contrast coefficients tables \cr
  \code{results$postHocs} \tab \tab \tab \tab \tab an array of post-hoc tables \cr
  \code{results$simpleEffects$Anova} \tab \tab \tab \tab \tab a table of ANOVA for simple effects \cr
  \code{results$simpleEffects$Params} \tab \tab \tab \tab \tab a table \cr
  \code{results$emeansTables} \tab \tab \tab \tab \tab an array of predicted means tables \cr
  \code{results$descPlot} \tab \tab \tab \tab \tab a descriptives plot \cr
  \code{results$descPlots} \tab \tab \tab \tab \tab an array of results plots \cr
  \code{results$plotnotes} \tab \tab \tab \tab \tab a html \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$info$asDF}

\code{as.data.frame(results$info)}
}
\description{
Generalized Mixed Linear Model
}
\examples{
data(subjects_by_stimuli)

gamlj::gamljMixed(
       formula = y ~ 1 + cond+( 1|subj ),
       data = subjects_by_stimuli)

gamlj::gamljMixed(
       data = subjects_by_stimuli,
       dep = "y",
       factors = "cond",
       modelTerms = "cond",
       cluster = "subj",
       randomTerms=list(list(c("cond","subj"))))

}
