% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamljgzlm.f.R
\name{gamljGzlm}
\alias{gamljGzlm}
\title{Generalized Linear Models}
\usage{
gamljGzlm(
  formula,
  data,
  dep = NULL,
  factors = NULL,
  covs = NULL,
  model_terms = NULL,
  nested_terms = NULL,
  fixed_intercept = TRUE,
  nested_intercept = TRUE,
  omnibus = "LRT",
  estimates_ci = TRUE,
  expb_ci = TRUE,
  es = list("expb"),
  model_type = "linear",
  custom_family = "gaussian",
  custom_link = "identity",
  propodds_test = FALSE,
  ci_width = 95,
  ci_method = "wald",
  boot_r = 1000,
  contrasts = NULL,
  show_contrastnames = TRUE,
  show_contrastcodes = FALSE,
  covs_scale = NULL,
  covs_conditioning = "mean_sd",
  ccm_value = 1,
  ccp_value = 25,
  covs_scale_labels = "labels",
  vcov = FALSE,
  emmeans = NULL,
  posthoc = NULL,
  adjust = list("bonf"),
  simple_effects = NULL,
  simple_moderators = NULL,
  simple_interactions = FALSE,
  plotHAxis = NULL,
  plotSepLines = NULL,
  plotSepPlots = NULL,
  plotRaw = FALSE,
  plotDvScale = FALSE,
  plotOriginalScale = FALSE,
  plotLinesTypes = FALSE,
  plotError = "none"
)
}
\arguments{
\item{formula}{(optional) the formula to use, see the examples}

\item{data}{the data as a data frame}

\item{dep}{a string naming the dependent variable from \code{data}; the
variable must be numeric. Not needed if \code{formula} is used.}

\item{factors}{a vector of strings naming the fixed factors from
\code{data}. Not needed if \code{formula} is used.}

\item{covs}{a vector of strings naming the covariates from \code{data}. Not
needed if \code{formula} is used.}

\item{model_terms}{a list of character vectors describing fixed effects
terms. Not needed if \code{formula} is used.}

\item{nested_terms}{a list of character vectors describing effects terms
for nestet. It can be passed as right-hand formula.}

\item{fixed_intercept}{\code{TRUE} (default) or \code{FALSE}, estimates
fixed intercept. Not needed if \code{formula} is used.}

\item{nested_intercept}{\code{TRUE} (default) or \code{FALSE}, estimates
fixed intercept. Not needed if \code{formula} is used.}

\item{omnibus}{Pseudo R-squared omnibus tests are based on
loglikelihood ratio test \code{'LRT'}  (default) or Wald's Chi-square \code{'wald'}.}

\item{estimates_ci}{\code{TRUE} (default) or \code{FALSE} , coefficients CI
in tables}

\item{expb_ci}{\code{TRUE} (default) or \code{FALSE} , exp(B) CI in table}

\item{es}{.}

\item{model_type}{Select the generalized linear model:
\code{linear},\code{poisson},\code{logistic},\code{multinomial}}

\item{custom_family}{Distribution family for the custom model, accepts
gaussian, binomial, gamma and inverse_gaussian .}

\item{custom_link}{Distribution family for the custom model, accepts
identity, log and inverse, onemu2 (for 1/mu^2).}

\item{propodds_test}{Test parallel lines assumptions in cumulative link
model (ordinal regression)}

\item{ci_width}{a number between 50 and 99.9 (default: 95) specifying the
confidence interval width for the plots.}

\item{ci_method}{.}

\item{boot_r}{a number bootstrap repetitions.}

\item{contrasts}{a named vector of the form \code{c(var1="type",
var2="type2")} specifying the type of contrast to use, one of
\code{'deviation'}, \code{'simple'}, \code{'dummy'}, \code{'difference'},
\code{'helmert'}, \code{'repeated'} or \code{'polynomial'}. If NULL,
\code{simple} is used. Can also be passed as a list of list of the form
list(list(var="var1",type="type1")).}

\item{show_contrastnames}{\code{TRUE} or \code{FALSE} (default), shows raw
names of the contrasts variables in tables}

\item{show_contrastcodes}{\code{TRUE} or \code{FALSE} (default), shows
contrast coefficients tables}

\item{covs_scale}{a named vector of the form \code{c(var1='type',
var2='type2')} specifying the transformation to apply to covariates, one of
\code{'centered'} to the mean, \code{'standardized'},\code{'log'} or
\code{'none'}. \code{'none'} leaves the variable as it is.}

\item{covs_conditioning}{\code{'mean_sd'} (default), or
\code{'percent'}. How to condition covariates in simple effects and plots.
\code{'mean_sd'} for mean +/- \code{ccp_value} * sd. Default is 1.
\code{'percent'} for median +/- \code{ccm_value} for percentiles. Default is 25.}

\item{ccm_value}{how many std around the means to condition
simple effects and plots. Used if \code{covs_conditioning}=\code{'mean_sd'}}

\item{ccp_value}{offsett (number of percentiles) around the median used to
condition simple effects and plots. Used if
\code{simpleScale}=\code{'percent'}}

\item{covs_scale_labels}{how the levels of a continuous moderator should
appear in tables and plots: \code{'labels'}, \code{'values'} and
\code{'values_labels'}, \code{'ovalues'}, \code{'ovalues_labels'}. The latter two refer
to the variable orginal levels, before scaling.}

\item{vcov}{\code{TRUE} or \code{FALSE} (default), shows coefficients
covariances}

\item{emmeans}{a rhs formula with the terms specifying the marginal means
to estimate (of the form \code{'~x+x:z'})}

\item{posthoc}{a rhs formula with the terms specifying the table to apply
the comparisons (of the form \code{'~x+x:z'}). The formula is not expanded,
so '\code{x*z}' becomes '\code{x+z' and not '}x+z+x:z\code{'. It can be
passed also as a list of the form '}list("x","z",c("x","z")`'}

\item{adjust}{one or more of \code{'none'},  \code{'bonf'},\code{'tukey'}
\code{'holm'}, \code{'scheffe'}; provide no,  Bonferroni, Tukey, Holm, Scheffe Post Hoc corrections
respectively.}

\item{simple_effects}{The variable for which the simple effects (slopes)
are computed}

\item{simple_moderators}{the variable that provides the levels at which the
simple effects are computed}

\item{simple_interactions}{should simple Interactions be computed}

\item{plotHAxis}{a string naming the variable placed on the horizontal axis
of the plot}

\item{plotSepLines}{a string naming the variable represented as separate
lines in the plot}

\item{plotSepPlots}{a list of string naming the variables defining the
levels for multiple plots}

\item{plotRaw}{\code{TRUE} or \code{FALSE} (default), plot raw data along
the predicted values}

\item{plotDvScale}{\code{TRUE} or \code{FALSE} (default), set the Y-axis
range equal to the range of the observed values.}

\item{plotOriginalScale}{\code{TRUE} or \code{FALSE} (default), use
original scale for covariates.}

\item{plotLinesTypes}{\code{TRUE} or \code{FALSE} (default), use different
linetypes per levels.}

\item{plotError}{\code{'none'} (default), \code{'ci'}, or \code{'se'}. Use
no error bars, use confidence intervals, or use standard errors on the
plots, respectively.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$model} \tab \tab \tab \tab \tab The underlying estimated model \cr
  \code{results$info} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$r2} \tab \tab \tab \tab \tab a table of R \cr
  \code{results$main$fit} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$anova} \tab \tab \tab \tab \tab a table of ANOVA results \cr
  \code{results$main$coefficients} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$vcov} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$contrastCodeTables} \tab \tab \tab \tab \tab an array of contrast coefficients tables \cr
  \code{results$main$marginals} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$relativerisk} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$paralleltest} \tab \tab \tab \tab \tab a table \cr
  \code{results$posthoc} \tab \tab \tab \tab \tab an array of post-hoc tables \cr
  \code{results$simpleEffects$anova} \tab \tab \tab \tab \tab a table of ANOVA for simple effects \cr
  \code{results$simpleEffects$coefficients} \tab \tab \tab \tab \tab a table \cr
  \code{results$simpleInteractions} \tab \tab \tab \tab \tab an array of simple interactions tables \cr
  \code{results$emmeans} \tab \tab \tab \tab \tab an array of predicted means tables \cr
  \code{results$mainPlots} \tab \tab \tab \tab \tab an array of results plots \cr
  \code{results$plotnotes} \tab \tab \tab \tab \tab a html \cr
  \code{results$predicted} \tab \tab \tab \tab \tab an output \cr
  \code{results$residuals} \tab \tab \tab \tab \tab an output \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$info$asDF}

\code{as.data.frame(results$info)}
}
\description{
Generalized Linear Model
}
\examples{
data<-emmeans::neuralgia
 gamlj::gamljGzlm(
           formula = Pain ~ Duration,
           data = data,
            model_type = "logistic")

}
