% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamljmixed.f.R
\name{gamljMixed}
\alias{gamljMixed}
\title{Mixed Model}
\usage{
gamljMixed(
  data,
  dep = NULL,
  factors = NULL,
  covs = NULL,
  modelTerms = NULL,
  fixedIntercept = TRUE,
  showParamsCI = TRUE,
  paramCIWidth = 95,
  ciRE = FALSE,
  contrasts = NULL,
  showRealNames = TRUE,
  showContrastCode = FALSE,
  plotHAxis = NULL,
  plotSepLines = NULL,
  plotSepPlots = NULL,
  plotRaw = FALSE,
  plotDvScale = FALSE,
  plotError = "none",
  ciWidth = 95,
  postHoc = NULL,
  eDesc = FALSE,
  eCovs = FALSE,
  simpleVariable = NULL,
  simpleModerator = NULL,
  simple3way = NULL,
  simpleScale = "mean_sd",
  cvalue = 1,
  percvalue = 25,
  simpleScaleLabels = "labels",
  postHocCorr = list("bonf"),
  scaling = NULL,
  dep_scale = "none",
  cluster = NULL,
  randomTerms = list(list()),
  correlatedEffects = "corr",
  reml = TRUE,
  lrtRandomEffects = FALSE,
  plotRandomEffects = FALSE,
  cimethod = "wald",
  dfmethod = "Satterthwaite",
  qq = FALSE,
  normTest = FALSE,
  normPlot = FALSE,
  residPlot = FALSE,
  clusterBoxplot = FALSE,
  randHist = FALSE,
  formula
)
}
\arguments{
\item{data}{the data as a data frame}

\item{dep}{a string naming the dependent variable from \code{data},
variable must be numeric}

\item{factors}{a vector of strings naming the fixed factors from
\code{data}}

\item{covs}{a vector of strings naming the covariates from \code{data}}

\item{modelTerms}{a list of character vectors describing fixed effects
terms}

\item{fixedIntercept}{\code{TRUE} (default) or \code{FALSE}, estimates
fixed intercept}

\item{showParamsCI}{\code{TRUE} (default) or \code{FALSE} , parameters CI
in table}

\item{paramCIWidth}{a number between 50 and 99.9 (default: 95) specifying
the confidence interval width for the parameter estimates}

\item{ciRE}{\code{TRUE} or \code{FALSE} (default), produce random effects confidence intervals based on profile method}

\item{contrasts}{a named vector of the form \code{c(var1='type', var2='type2')} specifying the type of contrast to use,
one of \code{'deviation'}, \code{'simple'}, \code{'dummy'}, \code{'difference'}, \code{'helmert'}, \code{'repeated'} or \code{'polynomial'}.
 If NULL, \code{'simple'} is used. Can also be passed as a list of list of the form \code{list(list(var='var1',type='type1'))}.}

\item{showRealNames}{\code{TRUE} or \code{FALSE} (default), provide raw
names of the contrasts variables}

\item{showContrastCode}{\code{TRUE} or \code{FALSE} (default), provide
contrast coefficients tables}

\item{plotHAxis}{a string naming the variable placed on the horizontal axis
of the plot}

\item{plotSepLines}{a string naming the variable represented as separate
lines on the plot}

\item{plotSepPlots}{the variable for whose levels multiple plots are
computed}

\item{plotRaw}{\code{TRUE} or \code{FALSE} (default), provide descriptive
statistics}

\item{plotDvScale}{\code{TRUE} or \code{FALSE} (default), scale the plot
Y-Axis to the max and the min of the dependent variable observed scores.}

\item{plotError}{\code{'none'}, \code{'ci'} (default), or \code{'se'}. Use
no error bars, use confidence intervals, or use standard errors on the
plots, respectively}

\item{ciWidth}{a number between 50 and 99.9 (default: 95) specifying the
confidence interval width}

\item{postHoc}{a list of terms to perform post-hoc tests on}

\item{eDesc}{\code{TRUE} or \code{FALSE} (default), provide lsmeans
statistics}

\item{eCovs}{\code{TRUE} or \code{FALSE} (default), provide lsmeans
statistics}

\item{simpleVariable}{The variable for which the simple effects (slopes)
are computed}

\item{simpleModerator}{the variable that provides the levels at which the
simple effects computed}

\item{simple3way}{a moderator of the two-way interaction which is probed}

\item{simpleScale}{\code{'mean_sd'} (default), \code{'custom'} , or
\code{'custom_percent'}. Use to condition the covariates (if any)}

\item{cvalue}{offset value for conditioning}

\item{percvalue}{offset value for conditioning}

\item{simpleScaleLabels}{decide the labeling of simple effects in tables
and plots.  \code{labels} indicates that only labels are used, such as
\code{Mean} and  \code{Mean + 1 SD}. \code{values} uses the actual values
as labels. \code{values_labels} uses both.}

\item{postHocCorr}{one or more of \code{'none'},  \code{'bonf'}, or
\code{'holm'}; provide no,  Bonferroni, and Holm Post Hoc corrections
respectively}

\item{scaling}{a named vector of the form \code{c(var1='type1',var2='type2')}. Types are
\code{'centered'} to the mean, \code{'clusterbasedcentered'} to the mean of each cluster, \code{'standardized'},
 \code{'clusterbasedstandardized'} standardized within each cluster, log-transformed \code{'log'},  or \code{'none'}.
\code{'none'} leaves the variable as it is. It can also be passed as a list of lists.}

\item{dep_scale}{Re-scale the dependent variable. It could be  \code{'centered'} to the mean, \code{'clusterbasedcentered'} to the mean of each cluster, \code{'standardized'},
\code{'clusterbasedstandardized'} standardized within each cluster, log-transformed \code{'log'},  or \code{'none'} (default).}

\item{cluster}{a vector of strings naming the clustering variables from
\code{data}}

\item{randomTerms}{a list of lists specifying the models random effects.}

\item{correlatedEffects}{\code{'nocorr'}, \code{'corr'} (default), or
\code{'block'}. When random effects are passed as list of length 1, it
decides whether the effects should be correlated,  non correlated. If
\code{'randomTerms'} is a list of  lists of length > 1, the option is
automatially set to \code{'block'}. The option is ignored if the model is
passed using \code{formula}.}

\item{reml}{\code{TRUE} (default) or \code{FALSE}, should the Restricted ML
be used rather than ML}

\item{lrtRandomEffects}{\code{TRUE} or \code{FALSE} (default), LRT for the
random effects}

\item{plotRandomEffects}{\code{TRUE} or \code{FALSE} (default), add random
effects predicted values in the plot}

\item{cimethod}{.}

\item{dfmethod}{.}

\item{qq}{\code{TRUE} or \code{FALSE} (default), provide a Q-Q plot of
residuals}

\item{normTest}{\code{TRUE} or \code{FALSE} (default), provide a test for
normality of residuals}

\item{normPlot}{\code{TRUE} or \code{FALSE} (default), provide a histogram
of residuals superimposed by a normal distribution}

\item{residPlot}{\code{TRUE} or \code{FALSE} (default), provide a
scatterplot of the residuals against predicted}

\item{clusterBoxplot}{\code{TRUE} or \code{FALSE} (default), provide a
boxplot of random effects by the first defined clustering variable}

\item{randHist}{\code{TRUE} or \code{FALSE} (default), provide histogram of
random Coefficients}

\item{formula}{(optional) the formula to use, see the examples#'}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$model} \tab \tab \tab \tab \tab The underlying \code{lm} object \cr
  \code{results$info} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$anova} \tab \tab \tab \tab \tab a table of ANOVA results \cr
  \code{results$main$fixed} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$random} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$randomCov} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$lrtRandomEffectsTable} \tab \tab \tab \tab \tab a table \cr
  \code{results$main$contrastCodeTables} \tab \tab \tab \tab \tab an array of contrast coefficients tables \cr
  \code{results$postHocs} \tab \tab \tab \tab \tab an array of post-hoc tables \cr
  \code{results$simpleEffects$Anova} \tab \tab \tab \tab \tab a table of ANOVA for simple effects \cr
  \code{results$simpleEffects$Params} \tab \tab \tab \tab \tab a table \cr
  \code{results$emeansTables} \tab \tab \tab \tab \tab an array of predicted means tables \cr
  \code{results$descPlot} \tab \tab \tab \tab \tab a descriptives plot \cr
  \code{results$descPlots} \tab \tab \tab \tab \tab an array of results plots \cr
  \code{results$plotnotes} \tab \tab \tab \tab \tab a html \cr
  \code{results$assumptions$normTest} \tab \tab \tab \tab \tab a table of normality tests \cr
  \code{results$assumptions$qq} \tab \tab \tab \tab \tab a q-q plot \cr
  \code{results$assumptions$normPlot} \tab \tab \tab \tab \tab Residual histogram \cr
  \code{results$assumptions$residPlot} \tab \tab \tab \tab \tab Residual Predicted plot \cr
  \code{results$assumptions$clusterBoxplot} \tab \tab \tab \tab \tab Residuals boxplot by cluster \cr
  \code{results$assumptions$randHist} \tab \tab \tab \tab \tab an array of random coefficients histograms \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$info$asDF}

\code{as.data.frame(results$info)}
}
\description{
Mixed Linear Model. Estimates models using `me4::lmer()` function and provides options to facilitate estimation of 
           interactions, simple slopes, simple effects, post-hoc tests, contrast analysis, effect size indexes and visualization of the results.
}
\examples{
data(subjects_by_stimuli)

gamlj::gamljMixed(
       formula = y ~ 1 + cond+( 1|subj ),
       data = subjects_by_stimuli)

gamlj::gamljMixed(
       data = subjects_by_stimuli,
       dep = "y",
       factors = "cond",
       modelTerms = "cond",
       cluster = "subj",
       randomTerms=list(list(c("cond","subj"))))
       
}
