% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{plot.gamljGlmResults}
\alias{plot.gamljGlmResults}
\title{Update a GAMLj model plot by passing new plot directives}
\usage{
\method{plot}{gamljGlmResults}(
  object,
  formula = NULL,
  haxis = NULL,
  sepLines = NULL,
  sepPlots = NULL,
  ...
)
}
\arguments{
\item{object}{a gamlj results object of the class `gamlj*Results`}

\item{formula}{a right hand side formula specifying the effect to plot, of the form `~x`, `~x*z` or `~x*z*w`. 
It can be combined with the other options but it has prevalence.}

\item{haxis}{horizontal axis variable}

\item{sepLines}{variable defining the levels for separate lines}

\item{sepPlots}{variable defining the levels for which separate plots are produced}

\item{...}{ignored}
}
\value{
an object of class GAMLj* as the input object
}
\description{
This function re-estimates a GAMLj model with a new plot. If not options is passed, extracts the 
plots present in the `gamlj*Results` object. If one plot is present, it is returned, if more than one is present, 
a list of plots is returned. An empty list is returned if no plot is present or defined.
}
\examples{
data(qsport)
mod<-gamlj::gamljGlm(
  formula = performance ~ hours,
  data = qsport)

plot(mod,haxis = "hours")
plot(mod,formula=~hours)
}
\author{
Marcello Gallucci
}
