% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{predict}
\alias{predict}
\alias{predict.gamljGlmResults}
\alias{predict.gamljGzlmResults}
\alias{predict.gamljMixedResults}
\alias{predict.gamljGlmMixedResults}
\title{Predicted values from GAMLj models}
\usage{
\method{predict}{gamljGlmResults}(object, ...)

\method{predict}{gamljGzlmResults}(object, type = "response", ...)

\method{predict}{gamljMixedResults}(object, re.form = NULL, type = "response", ...)

\method{predict}{gamljGlmMixedResults}(object, re.form = NULL, type = "response", ...)
}
\arguments{
\item{object}{a gamlj results object of the class `gamlj*Results`}

\item{...}{additional arguments for specific predict methods other than the ones specified here.}

\item{type}{the type of prediction required. The default is on the scale of the response variables ('response'); Thus for  binomial models the default is to compute the predicted probabilities.  'link' gives the scale of the linear predictors; 
is on the scale of the linear predictors;  The 'terms' option returns a matrix giving the fitted values of each term in the model formula on the linear predictor scale. 
Cf. \code{\link[stats:predict]{stats::predict()}}, \code{\link[stats:predict.lm]{stats::predict.lm()}}}

\item{re.form}{(formula, NULL, or NA) specify which random effects to condition on when predicting. If NULL, include all random effects; if NA or ~0, include no random effects. Used only for the mixed models.}
}
\value{
a R object of the class of the estimated model
}
\description{
Returns predicted values from the estimated model
}
\examples{
data('qsport')
gmod<-gamlj::gamljGlm(
   formula = performance ~ hours,
   data = qsport)
 preds<-predict(gmod)
 
}
\author{
Marcello Gallucci
}
