% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{summary.ResultsElement}
\alias{summary.ResultsElement}
\alias{summary.gamlj_list}
\alias{print.jmvrtable}
\alias{print.jmvrobj}
\alias{print.gamlj_list}
\alias{coef.gamlj}
\alias{fit}
\alias{fit.gamlj}
\alias{em_means.gamlj}
\title{S3 methods for class jamovi ResultsElement}
\usage{
\method{summary}{ResultsElement}(object, ...)

\method{summary}{gamlj_list}(object, ...)

\method{print}{jmvrtable}(x, ...)

\method{print}{jmvrobj}(x, ...)

\method{print}{gamlj_list}(x, ...)

\method{coef}{gamlj}(object, ...)

fit(x, ...)

\method{fit}{gamlj}(x, ...)

\method{em_means}{gamlj}(object, formula = NULL, ...)
}
\arguments{
\item{object}{a gamlj results object of the class `gamlj`}

\item{...}{additional arguments passed to the GAMLj estimation function}

\item{x}{a gamlj results object of the class `gamlj`}
}
\value{
a list of table as data.frame

a list of table as data.frame
}
\description{
These functions extract all visible tables from a ResultsElement or related classes produced by gamlj
and print them in R style.

These functions extract all visible tables from a a list of table produced by gamlj
and print them in R style.
}
\examples{
data(fivegroups)
fivegroups$Group<-factor(fivegroups$Group)
gmod<-GAMLj3::gamlj_lm(
  formula = Score ~Group,
  data = fivegroups)

summary(gmod)
data(fivegroups)
fivegroups$Group<-factor(fivegroups$Group)
gmod<-GAMLj3::gamlj_lm(
  formula = Score ~Group,
  data = fivegroups)

summary(gmod)
}
\author{
Marcello Gallucci
}
