% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temptime}
\alias{temptime}
\title{Individual growth model data}
\usage{
data(temptime)
}
\description{
Data for Individual growth model data examples
}
\examples{
data(temptime)
}
\keyword{datasets}
