% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{update.gamljGlmResults}
\alias{update.gamljGlmResults}
\alias{update.gamljGzlmResults}
\alias{update.gamljMixedResults}
\alias{update.gamljGlmMixedResults}
\title{Update a GAMLj model by passing only the new options}
\usage{
\method{update}{gamljGlmResults}(object, ...)

\method{update}{gamljGzlmResults}(object, ...)

\method{update}{gamljMixedResults}(object, ...)

\method{update}{gamljGlmMixedResults}(object, ...)
}
\arguments{
\item{object}{of class `gamlj*Results`}

\item{...}{any parameter to be passed to \code{\link{gamljGlm}},  \code{\link{gamljMixed}},  \code{\link{gamljGzlm}}, or  \code{\link{gamljGlmMixed}}}
}
\value{
an object of class gamlj*Results as the input object
}
\description{
Re-estimates a GAMLj model applying new options to the original model
}
\author{
Marcello Gallucci
}
