% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{update.gamljlmResults}
\alias{update.gamljlmResults}
\alias{update.gamljglmResults}
\alias{update.gamljmixedResults}
\alias{update.gamljgmixedResults}
\title{Update a GAMLj model}
\usage{
update.gamljlmResults(object, ...)

update.gamljglmResults(object, ...)

update.gamljmixedResults(object, ...)

update.gamljgmixedResults(object, ...)
}
\arguments{
\item{object}{of class `gamlj*Results`}

\item{...}{any parameter to be passed to \code{\link[GAMLj3]{gamlj_lm}},  \code{\link[GAMLj3]{gamlj_mixed}},  \code{\link[GAMLj3]{gamlj_glm}}, or  \code{\link{gamlj_gmixed}}}
}
\value{
an object of class gamlj*Results as the input object
}
\description{
Re-estimates a GAMLj model applying new options to the original model
}
\author{
Marcello Gallucci
}
