% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{vcov.gamlj}
\alias{vcov.gamlj}
\title{Coefficients covariances for a GAMLj model}
\usage{
\method{vcov}{gamlj}(object, ...)
}
\arguments{
\item{object}{of class `gamlj*Results`}

\item{...}{any parameter to be passed to \code{\link[gamlj]{gamlj_lm}},  \code{\link[gamlj]{gamlj_mixed}},  \code{\link[gamlj]{gamlj_glm}}, or  \code{\link[gamlj]{gamlj_gmixed}}}
}
\value{
an table of class gamlj*Results
}
\description{
Returns the parameters variances and covariances
}
\author{
Marcello Gallucci
}
